CyTOF data from two samples: 5,000 B-cell lineage cells from a healthy patient and 5,000 B-cell lineage cells from a B-cell precursor Acute Lymphoblastic Leukemia (BCP-ALL) patient.

A dataset containing CyTOF measurements from immune cells originally studied in the following paper:

Good Z, Sarno J, et al. Single-cell developmental classification of B cell precursor acute lymphoblastic leukemia at diagnosis reveals predictors of relapse. Nat Med. 2018 May;24(4):474-483. doi: 10.1038/nm.4505. Epub 2018 Mar 5. PMID: 29505032; PMCID: PMC5953207.

Usage

data(ddpr_data)

Format

A data frame with 10000 rows and 24 variables:

sample_name

name of the sample from which the data was read

cd45

A CyTOF measurement in raw ion counts

cd19

A CyTOF measurement in raw ion counts

cd22

A CyTOF measurement in raw ion counts

cd79b

A CyTOF measurement in raw ion counts

cd20

A CyTOF measurement in raw ion counts

cd34

A CyTOF measurement in raw ion counts

cd123

A CyTOF measurement in raw ion counts

cd10

A CyTOF measurement in raw ion counts

cd24

A CyTOF measurement in raw ion counts

cd127

A CyTOF measurement in raw ion counts

cd43

A CyTOF measurement in raw ion counts

cd38

A CyTOF measurement in raw ion counts

cd58

A CyTOF measurement in raw ion counts

psyk

A CyTOF measurement in raw ion counts

p4ebp1

A CyTOF measurement in raw ion counts

pstat5

A CyTOF measurement in raw ion counts

pakt

A CyTOF measurement in raw ion counts

ps6

A CyTOF measurement in raw ion counts

perk

A CyTOF measurement in raw ion counts

pcreb

A CyTOF measurement in raw ion counts

Source

https://github.com/kara-davis-lab/DDPR

Value

A data.frame