Clinical metadata for each patient sample in Good & Sarno et al. (2018).

A dataset containing patient-level clinical metadata for samples originally studied in the following paper:

Good Z, Sarno J, et al. Single-cell developmental classification of B cell precursor acute lymphoblastic leukemia at diagnosis reveals predictors of relapse. Nat Med. 2018 May;24(4):474-483. doi: 10.1038/nm.4505. Epub 2018 Mar 5. PMID: 29505032; PMCID: PMC5953207.

Usage

data(ddpr_metadata)

Format

A data frame with 10000 rows and 12 variables:

patient_id

Name of the sample from which the data was read

gender

Gender of the patient from which each sample was collected

age_at_diagnosis

Age (in years) of the patient from which each sample was collected

wbc_count

The diagnostic White Blood Cell (WBC) count of the patient from which each sample was collected

mrd_risk

Risk stratification category for each patient using minimal residual disease (MRD) criteria

nci_rome_risk

Risk stratification category for each patient using National Cancer Institute (NCI) criteria

relapse_status

A string representing whether or not a patient relapsed

time_to_relapse

The time (in days) it took each patient to relapse. Patients who did not relapse will have the value of NA

type_of_relapse

A string representing the timing of relapse for each patient. "Very early" relapses occurred less than 18 months after diagnosis; "Early" relapses occurred between 18 months and 32 months after diagnosis; "Late" relapses occurred later than 32 months after diagnosis.

ccr

The number of documented days of continuous complete remission (CCR) for patients who did not relapse. All patients who relapsed will have a value of NA.

cohort

A string representing if each sample was used in the "Training" or "Validation" cohort in the original study

ddpr_risk

The risk category ("Low" or "High") assigned to each sample using the original paper's risk-stratification algorithm

Source

Good Z, Sarno J, et al. Single-cell developmental classification of B cell precursor acute lymphoblastic leukemia at diagnosis reveals predictors of relapse. Nat Med. 2018 May;24(4):474-483. doi: 10.1038/nm.4505. Epub 2018 Mar 5. PMID: 29505032; PMCID: PMC5953207. Supplementary Table 1.

Value

A data.frame