Constructor for a tof_model.

Constructor for a tof_model.

Usage

new_tof_model(
  model,
  recipe,
  penalty,
  mixture,
  model_type = c("linear", "two-class", "multiclass", "survival"),
  outcome_colnames,
  training_data
)

Arguments

model

A glmnet model.

recipe

A prepped recipe object.

penalty

A double indicating which lambda value should be used within the glmnet path.

mixture

A double indicating which alpha value was used to fit the glmnet model.

model_type

A string indicating which type of glmnet model is being fit.

outcome_colnames

TO DO

training_data

TO DO

Value

A `tof_model`, an S3 class that includes a trained glmnet model and the recipe used to perform its associated preprocessing.