Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>       cd45    cd38   cd34   cd19 cluster_id
#>      <dbl>   <dbl>  <dbl>  <dbl> <chr>     
#>  1 -0.417  -0.855   1.16  -0.347 n         
#>  2  0.224   2.11    1.35   0.121 r         
#>  3 -2.25   -2.44   -0.100 -1.46  f         
#>  4 -0.540   0.0877 -0.326  0.220 w         
#>  5  0.108  -0.0184  0.644  0.836 o         
#>  6 -0.436  -0.0671  0.765  1.16  a         
#>  7 -0.0216  0.554  -0.330 -0.404 p         
#>  8 -1.23    0.521   1.20   1.43  p         
#>  9  0.528  -2.25   -0.658 -0.764 x         
#> 10 -0.157  -0.625  -0.694  0.101 q         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>        cd45   cd38   cd34    cd19 cluster_id
#>       <dbl>  <dbl>  <dbl>   <dbl> <chr>     
#>  1 -0.202   -1.30   0.945  1.20   i         
#>  2 -1.97     0.313 -1.12  -0.105  w         
#>  3 -0.614    1.43   0.530  0.219  i         
#>  4 -0.907    0.823  0.249  0.0425 a         
#>  5 -0.0517  -0.822  0.954 -0.335  r         
#>  6  1.45     0.749  0.161  0.702  v         
#>  7  0.171    1.20  -1.09  -0.134  p         
#>  8  0.00567 -1.82  -0.810 -1.01   e         
#>  9  0.503   -0.373 -0.494  0.342  r         
#> 10 -0.409    0.174 -0.393 -0.155  d         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 104 × 5
#>        cd45    cd38   cd34   cd19 cluster_id
#>       <dbl>   <dbl>  <dbl>  <dbl> <chr>     
#>  1  1.75    -1.14    0.247  1.68  m         
#>  2  0.224    2.11    1.35   0.121 r         
#>  3 -2.25    -2.44   -0.100 -1.46  f         
#>  4 -0.115   -1.02   -0.537 -0.926 p         
#>  5  1.86     1.09   -1.07  -0.803 s         
#>  6  0.00101 -0.777  -2.06   1.77  o         
#>  7 -1.20     1.35   -0.228 -1.31  y         
#>  8  0.274   -0.597  -1.02   1.98  m         
#>  9  0.692    0.951   0.347  0.129 n         
#> 10 -1.18    -0.0734  0.846 -0.866 u         
#> # ℹ 94 more rows