Downsample high-dimensional cytometry data by randomly selecting a constant number of cells per group.

This function downsamples the number of cells in a `tof_tbl` by randomly selecting `num_cells` cells from each unique combination of values in `group_cols`.

Usage

tof_downsample_constant(tof_tibble, group_cols = NULL, num_cells)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups from which `num_cells` will be downsampled. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

num_cells

An integer number of cells that should be sampled from each group defined by `group_cols`.

Value

A `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. Specifically, the number of rows will be `num_cells` multiplied by the number of unique combinations of the values in `group_cols`. If any group has fewer than `num_cells` number of cells, all cells from that group will be kept.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 500 cells from the input data
tof_downsample_constant(
    tof_tibble = sim_data,
    num_cells = 500L
)
#> # A tibble: 500 × 5
#>      cd45    cd38   cd34   cd19 cluster_id
#>     <dbl>   <dbl>  <dbl>  <dbl> <chr>     
#>  1  0.264 -0.886  -0.650 -1.29  q         
#>  2  0.385  0.539  -0.534 -0.925 t         
#>  3 -0.460  1.22   -0.830 -0.734 i         
#>  4  1.24  -0.796  -0.926  0.228 f         
#>  5  1.79  -2.38    0.998  1.19  y         
#>  6 -0.374 -0.112  -0.579  2.84  k         
#>  7 -1.76  -1.01    1.03  -0.574 m         
#>  8 -1.99   2.16   -0.780 -0.475 u         
#>  9 -0.877  1.28    0.703  0.432 g         
#> 10 -0.256 -0.0268  0.750 -0.139 c         
#> # ℹ 490 more rows

# sample 20 cells per cluster from the input data
tof_downsample_constant(
    tof_tibble = sim_data,
    group_cols = cluster_id,
    num_cells = 20L
)
#> # A tibble: 520 × 5
#>       cd45   cd38   cd34   cd19 cluster_id
#>      <dbl>  <dbl>  <dbl>  <dbl> <chr>     
#>  1  0.264  -0.886 -0.650 -1.29  q         
#>  2 -0.460   1.22  -0.830 -0.734 i         
#>  3  1.24   -0.796 -0.926  0.228 f         
#>  4  1.06    0.770  0.543  1.74  j         
#>  5 -1.58   -1.21  -0.449 -0.520 m         
#>  6 -0.877   1.28   0.703  0.432 g         
#>  7  0.209  -0.129  0.749  0.846 n         
#>  8  0.539  -0.197  2.76  -0.676 s         
#>  9 -0.110   0.621 -0.109 -0.236 a         
#> 10 -0.0957 -2.46   0.493  0.446 t         
#> # ℹ 510 more rows