Downsample high-dimensional cytometry data by randomly selecting a proportion of the cells in each group.

This function downsamples the number of cells in a `tof_tbl` by randomly selecting a `prop_cells` proportion of the total number of cells with each unique combination of values in `group_cols`.

Usage

tof_downsample_prop(tof_tibble, group_cols = NULL, prop_cells)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups from which `prop_cells` will be downsampled. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

prop_cells

A proportion of cells (between 0 and 1) that should be sampled from each group defined by `group_cols`.

Value

A `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. Specifically, the number of rows should be `prop_cells` times the number of rows in the input `tof_tibble`.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 10% of all cells from the input data
tof_downsample_prop(
    tof_tibble = sim_data,
    prop_cells = 0.1
)
#> # A tibble: 100 × 5
#>       cd45    cd38    cd34    cd19 cluster_id
#>      <dbl>   <dbl>   <dbl>   <dbl> <chr>     
#>  1 -0.172  -0.338  -2.25    0.0232 n         
#>  2  0.393   0.828  -0.671   0.786  r         
#>  3 -1.30   -0.0369  0.680   1.34   r         
#>  4 -0.814   0.194  -1.19    0.851  z         
#>  5 -0.661   0.0125  0.224   1.31   a         
#>  6 -0.379  -0.836   0.701  -0.779  a         
#>  7 -0.0480 -0.812  -0.0560 -0.0699 k         
#>  8  0.599  -0.134   0.398  -2.00   v         
#>  9 -0.686  -1.57   -0.820  -0.280  x         
#> 10 -0.461   0.189  -0.362   0.0698 k         
#> # ℹ 90 more rows

# sample 10% of all cells from each cluster in the input data
tof_downsample_prop(
    tof_tibble = sim_data,
    group_cols = cluster_id,
    prop_cells = 0.1
)
#> # A tibble: 91 × 5
#>       cd45    cd38     cd34   cd19 cluster_id
#>      <dbl>   <dbl>    <dbl>  <dbl> <chr>     
#>  1  1.51   -1.06    1.09    -1.65  a         
#>  2  0.881   2.22   -0.262   -0.478 a         
#>  3  0.558   2.39   -0.334   -0.131 a         
#>  4 -0.628   1.51    1.51    -0.198 b         
#>  5 -0.499  -0.698   1.05    -0.269 b         
#>  6 -0.457   1.27    0.908   -0.887 b         
#>  7  0.162  -0.254  -0.0384  -0.636 b         
#>  8  1.40   -2.49   -1.97    -0.686 b         
#>  9  0.535  -0.0699  0.895    0.851 c         
#> 10 -0.0122  0.732  -0.00838  1.17  c         
#> # ℹ 81 more rows