Metacluster clustered CyTOF data.

This function is a wrapper around tidytof's tof_metacluster_* function family. It performs metaclustering on CyTOF data using a user-specified method (of 5 choices) and each method's corresponding input parameters.

Usage

tof_metacluster(
  tof_tibble,
  cluster_col,
  metacluster_cols = where(tof_is_numeric),
  central_tendency_function = stats::median,
  ...,
  augment = TRUE,
  method = c("consensus", "hierarchical", "kmeans", "phenograph", "flowsom")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

cluster_col

An unquoted column name indicating which column in `tof_tibble` stores the cluster ids for the cluster to which each cell belongs. Cluster labels can be produced via any method the user chooses - including manual gating, any of the functions in the `tof_cluster_*` function family, or any other method.

metacluster_cols

Unquoted column names indicating which columns in `tof_tibble` to use in computing the metaclusters. Defaults to all numeric columns in `tof_tibble`. Supports tidyselect helpers.

central_tendency_function

The function that should be used to calculate the measurement of central tendency for each cluster before metaclustering. This function will be used to compute a summary statistic for each input cluster in `cluster_col` across all columns specified by `metacluster_cols`, and the resulting vector (one for each cluster) will be used as the input for metaclustering. Defaults to median.

...

Additional arguments to pass to the `tof_metacluster_*` function family member corresponding to the chosen `method`.

augment

A boolean value indicating if the output should column-bind the metacluster ids of each cell as a new column in `tof_tibble` (TRUE; the default) or if a single-column tibble including only the metacluster ids should be returned (FALSE).

method

A string indicating which clustering method should be used. Valid values include "consensus", "hierarchical", "kmeans", "phenograph", and "flowsom".

Value

A `tof_tbl` or `tibble` If augment = FALSE, it will have a single column encoding the metacluster ids for each cell in `tof_tibble`. If augment = TRUE, it will have ncol(tof_tibble) + 1 columns: each of the (unaltered) columns in `tof_tibble` plus an additional column encoding the metacluster ids.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

tof_metacluster(
    tof_tibble = sim_data,
    cluster_col = cluster_id,
    clustering_algorithm = "consensus",
    method = "flowsom"
)
#> # A tibble: 1,000 × 6
#>       cd45    cd38   cd34     cd19 cluster_id .flowsom_metacluster
#>      <dbl>   <dbl>  <dbl>    <dbl> <chr>      <chr>               
#>  1  0.526   0.809  -1.18  -0.659   d          2                   
#>  2 -1.46   -1.45   -0.988 -1.58    s          1                   
#>  3  0.282  -1.10    0.911  1.17    v          2                   
#>  4  1.12    0.659  -1.20  -1.07    k          2                   
#>  5 -0.198  -0.0439 -1.04  -2.25    i          3                   
#>  6  1.27    0.855  -2.73   0.00435 j          3                   
#>  7  0.0878 -0.239   0.774  0.214   q          1                   
#>  8  1.04   -1.60    1.07   1.62    i          3                   
#>  9  1.26    0.411  -0.568 -0.712   q          1                   
#> 10  0.0907 -0.167  -0.566  0.114   a          1                   
#> # ℹ 990 more rows

tof_metacluster(
    tof_tibble = sim_data,
    cluster_col = cluster_id,
    method = "phenograph"
)
#> # A tibble: 1,000 × 6
#>       cd45    cd38   cd34     cd19 cluster_id .phenograph_metacluster
#>      <dbl>   <dbl>  <dbl>    <dbl> <chr>      <chr>                  
#>  1  0.526   0.809  -1.18  -0.659   d          1                      
#>  2 -1.46   -1.45   -0.988 -1.58    s          4                      
#>  3  0.282  -1.10    0.911  1.17    v          1                      
#>  4  1.12    0.659  -1.20  -1.07    k          1                      
#>  5 -0.198  -0.0439 -1.04  -2.25    i          1                      
#>  6  1.27    0.855  -2.73   0.00435 j          2                      
#>  7  0.0878 -0.239   0.774  0.214   q          2                      
#>  8  1.04   -1.60    1.07   1.62    i          1                      
#>  9  1.26    0.411  -0.568 -0.712   q          2                      
#> 10  0.0907 -0.167  -0.566  0.114   a          2                      
#> # ℹ 990 more rows