Train a recipe or list of recipes for preprocessing sample-level cytometry data

Train a recipe or list of recipes for preprocessing sample-level cytometry data

Usage

tof_prep_recipe(split_data, unprepped_recipe)

Arguments

split_data

An `rsplit` or `rset` object from the rsample package containing the sample-level data to use for modeling. The easiest way to generate this is to use tof_split_data. Alternatively, an unsplit tbl_df, though this is not recommended.

unprepped_recipe

A recipe object (if `split_data` is an `rsplit` object or a `tbl_df`) or list of recipes (if `split_data` is an `rset` object).

Value

If split_data is an "rsplit" or "tbl_df" object, will return a single prepped recipe. If split_data is an "rset" object, will return a list of prepped recipes specific for each fold of the resampling procedure.