Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>       cd45   cd38   cd34   cd19    .pc1   .pc2     .pc3    .pc4
#>      <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>    <dbl>   <dbl>
#>  1 -0.561  -0.259  2.01   0.845  1.12    0.283  1.46     1.16  
#>  2  0.332  -1.31   1.39   0.389  1.53    0.121  0.00464  0.920 
#>  3 -2.33    1.63   0.112 -0.119 -2.14    0.519  1.72     0.914 
#>  4 -1.18    1.42   0.597 -0.744 -1.48   -0.628  1.14     0.854 
#>  5  0.666   0.469 -0.181 -0.449 -0.306  -0.736 -0.256   -0.428 
#>  6  1.70   -1.30   1.42   0.239  2.22   -0.718 -0.496    0.184 
#>  7 -0.474   0.138 -1.46   0.995 -0.717   1.50   0.0360  -0.949 
#>  8 -0.0121 -1.38  -0.248  0.624  0.656   1.11  -0.614    0.183 
#>  9  0.728  -1.33  -0.111  0.147  0.870   0.315 -1.00     0.0103
#> 10  1.03    0.181 -0.171 -0.463  0.0189 -0.830 -0.546   -0.520 
#> # ℹ 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>       cd45   cd38   cd34   cd19 .tsne1  .tsne2
#>      <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1 -0.561  -0.259  2.01   0.845  -1.12 -3.02  
#>  2  0.332  -1.31   1.39   0.389  -3.26 -1.49  
#>  3 -2.33    1.63   0.112 -0.119   5.46 -0.0530
#>  4 -1.18    1.42   0.597 -0.744   4.78  0.0140
#>  5  0.666   0.469 -0.181 -0.449   2.38 -1.80  
#>  6  1.70   -1.30   1.42   0.239  -3.61 -2.86  
#>  7 -0.474   0.138 -1.46   0.995   1.38  2.17  
#>  8 -0.0121 -1.38  -0.248  0.624  -3.08  0.895 
#>  9  0.728  -1.33  -0.111  0.147  -3.65  0.923 
#> 10  1.03    0.181 -0.171 -0.463   2.36 -2.19  
#> # ℹ 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> # A tibble: 100 × 6
#>       cd45   cd38   cd34   cd19 .umap1 .umap2
#>      <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 -0.561  -0.259  2.01   0.845 -0.133 -0.963
#>  2  0.332  -1.31   1.39   0.389 -4.09   0.285
#>  3 -2.33    1.63   0.112 -0.119  4.83  -1.72 
#>  4 -1.18    1.42   0.597 -0.744  4.61  -1.28 
#>  5  0.666   0.469 -0.181 -0.449  2.49   1.01 
#>  6  1.70   -1.30   1.42   0.239 -3.36   0.151
#>  7 -0.474   0.138 -1.46   0.995 -0.808 -2.59 
#>  8 -0.0121 -1.38  -0.248  0.624 -5.92  -0.166
#>  9  0.728  -1.33  -0.111  0.147 -4.87   0.746
#> 10  1.03    0.181 -0.171 -0.463  2.25   0.798
#> # ℹ 90 more rows