Write cytof data to a file or to a directory of files

Write data (in the form of a `tof_tbl`) into either a .csv or an .fcs file for storage.

Usage

tof_write_data(
  tof_tibble = NULL,
  group_cols,
  out_path = NULL,
  format = c("fcs", "csv"),
  sep = "_",
  file_name
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Optional. Unquoted names of the columns in `tof_tibble` that should be used to group cells into separate files. Supports tidyselect helpers. Defaults to no grouping (all cells are written into a single file).

out_path

Path to the directory where output files should be saved.

format

format for the files being written. Currently supports .csv and .fcs files

sep

Delimiter that should be used between each of the values of `group_cols` to create the output .csv/.fcs file names. Defaults to "_".

file_name

If `group_cols` isn't specified, the name (without an extension) that should be used for the saved file.

Value

This function does not explicitly return any values. Instead, it writes .csv and/or .fcs files to the specified `out_path`.

See also