Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>      cd45    cd38   cd34    cd19 cluster_id
#>     <dbl>   <dbl>  <dbl>   <dbl> <chr>     
#>  1  0.124 -0.0156 -2.05   0.0134 g         
#>  2  1.65  -1.39   -0.180 -0.710  c         
#>  3  0.636 -0.170   0.778 -1.97   a         
#>  4 -1.44   0.340   0.241  1.63   n         
#>  5 -0.876 -1.04    0.109  0.403  i         
#>  6 -2.78  -0.871  -0.725  0.717  d         
#>  7  0.809 -0.349  -0.243 -0.735  w         
#>  8 -1.10  -0.263  -2.10  -1.32   j         
#>  9  1.95  -0.584  -0.993 -0.651  p         
#> 10  1.99  -0.220  -1.03  -2.58   a         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>      cd45   cd38   cd34    cd19 cluster_id
#>     <dbl>  <dbl>  <dbl>   <dbl> <chr>     
#>  1  0.797  0.755  1.77   0.876  r         
#>  2 -0.570  0.751 -1.00  -0.810  n         
#>  3 -0.450 -0.998 -0.986  1.78   w         
#>  4 -0.311  1.98  -0.719 -1.06   z         
#>  5 -0.669 -0.263  0.830  1.43   f         
#>  6  0.661 -0.696 -0.728 -0.639  l         
#>  7 -0.165  0.176 -0.203  0.0940 f         
#>  8 -0.770  0.292  0.309  0.132  t         
#>  9  0.490  1.07  -0.408 -0.541  a         
#> 10  0.874  0.445  1.02  -0.806  u         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 101 × 5
#>      cd45   cd38     cd34    cd19 cluster_id
#>     <dbl>  <dbl>    <dbl>   <dbl> <chr>     
#>  1  2.02  -0.234  1.02    -0.727  i         
#>  2  1.52  -0.652  2.83    -0.0350 q         
#>  3  0.281 -0.616 -0.00816 -1.07   k         
#>  4  1.79  -0.490  0.426    1.61   y         
#>  5  1.99  -0.220 -1.03    -2.58   a         
#>  6  1.43   1.35   0.391   -0.886  x         
#>  7 -0.335  1.80  -0.470   -0.747  r         
#>  8 -1.03  -0.145 -1.36    -1.08   c         
#>  9  0.630 -0.491 -0.990   -1.11   z         
#> 10 -0.110  0.501  0.903   -1.36   n         
#> # ℹ 91 more rows