Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>      cd45    cd38    cd34    cd19 cluster_id
#>     <dbl>   <dbl>   <dbl>   <dbl> <chr>     
#>  1  1.26   0.998   0.444  -0.591  h         
#>  2  0.341  1.50   -0.0357  1.11   a         
#>  3  1.39   0.0528  0.900   0.851  o         
#>  4 -0.542 -1.44    0.510   0.435  a         
#>  5  1.14   1.77   -2.72    2.04   p         
#>  6 -0.271 -0.450   0.541   0.325  k         
#>  7 -1.19  -0.680   0.459   0.557  b         
#>  8 -0.492  0.726   0.215  -1.06   g         
#>  9 -0.926  1.10    0.989   0.741  s         
#> 10 -1.31   0.203   1.83   -0.0195 c         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>      cd45    cd38    cd34   cd19 cluster_id
#>     <dbl>   <dbl>   <dbl>  <dbl> <chr>     
#>  1 -0.271 -0.450   0.541   0.325 k         
#>  2  1.49  -0.0481  1.25    0.344 q         
#>  3 -0.190 -1.77   -0.665   0.240 n         
#>  4 -1.25  -0.0155  0.439  -1.53  r         
#>  5 -1.05   0.616  -0.159   0.989 m         
#>  6 -0.236  1.32   -0.388   0.134 i         
#>  7 -1.22   1.36    0.209   0.293 m         
#>  8 -1.50  -0.976  -2.46    0.688 r         
#>  9  0.287  0.0766 -0.0885  0.133 l         
#> 10 -1.34  -0.197  -1.06    1.45  f         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 93 × 5
#>      cd45    cd38     cd34   cd19 cluster_id
#>     <dbl>   <dbl>    <dbl>  <dbl> <chr>     
#>  1 -0.754 -1.28    0.537    0.824 z         
#>  2 -1.52   0.146   0.895   -1.11  v         
#>  3 -0.952  0.705  -0.481   -1.59  p         
#>  4  0.668  0.911  -0.406    1.03  e         
#>  5 -1.58   0.504   1.99    -0.764 r         
#>  6  1.22   2.31   -0.486    0.649 a         
#>  7  0.417 -2.14   -1.70     0.713 y         
#>  8  1.36   0.199  -0.00351  1.11  b         
#>  9 -0.546  0.574   1.65     0.230 r         
#> 10  0.658 -0.0507  1.25     0.855 c         
#> # ℹ 83 more rows