Downsample CyTOF data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>      cd45    cd38   cd34    cd19 cluster_id
#>     <dbl>   <dbl>  <dbl>   <dbl> <chr>     
#>  1 -0.751  0.176  -0.163  0.0145 k         
#>  2 -0.936 -0.0406  0.459 -0.0936 x         
#>  3 -0.913 -0.266   0.502  0.0926 l         
#>  4 -0.306  1.21    0.189  1.37   e         
#>  5  0.203  0.279   1.47  -1.77   s         
#>  6  0.759 -1.30    1.06  -0.917  q         
#>  7  0.357 -1.19    0.403 -0.785  e         
#>  8 -1.72  -0.544  -0.452  1.98   i         
#>  9  0.183  1.20   -1.11  -0.897  u         
#> 10 -1.43   0.782  -1.75  -0.299  p         
#> # … with 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>      cd45   cd38    cd34   cd19 cluster_id
#>     <dbl>  <dbl>   <dbl>  <dbl> <chr>     
#>  1 -0.310 -1.02  -0.611  -1.15  d         
#>  2  0.652  0.961 -0.655   1.41  t         
#>  3  0.409 -1.40  -0.894   0.703 s         
#>  4 -0.536 -0.411  2.25    1.31  j         
#>  5 -0.418 -1.11   1.19   -0.807 c         
#>  6 -0.398 -0.638  0.839   0.689 s         
#>  7  0.940 -0.143 -1.09   -0.171 u         
#>  8 -0.199  0.191  0.441   1.29  c         
#>  9 -0.328 -0.140  0.0824  0.427 e         
#> 10  0.702  0.321  0.449  -1.49  y         
#> # … with 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 98 × 5
#>      cd45   cd38   cd34   cd19 cluster_id
#>     <dbl>  <dbl>  <dbl>  <dbl> <chr>     
#>  1 -1.49   1.21  -1.57   0.629 i         
#>  2 -0.306  1.21   0.189  1.37  e         
#>  3 -1.52  -0.884 -0.940  0.194 w         
#>  4  0.616 -0.330 -0.684  0.713 s         
#>  5 -0.199 -0.221 -1.21  -1.67  s         
#>  6  0.442 -0.664  1.02   1.10  l         
#>  7 -1.89   1.17   0.403 -0.202 v         
#>  8  1.25   0.345  0.155 -0.386 b         
#>  9 -0.944 -0.177  0.399 -0.851 q         
#> 10 -1.19  -0.302 -1.21   1.12  n         
#> # … with 88 more rows