Downsample CyTOF data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>      cd45   cd38    cd34   cd19 cluster_id
#>     <dbl>  <dbl>   <dbl>  <dbl> <chr>     
#>  1  0.174  1.44   2.69    1.62  z         
#>  2  0.433 -0.906 -0.411  -2.57  p         
#>  3  0.415 -0.425 -0.852  -1.46  z         
#>  4 -1.39  -0.663 -0.0319  0.557 n         
#>  5  0.847  0.330  0.419   0.745 j         
#>  6  0.136 -1.32  -0.926  -1.28  r         
#>  7  0.837 -0.541  1.38   -1.04  w         
#>  8  1.82  -0.120 -1.95    0.991 c         
#>  9 -0.382  0.283 -0.586   0.139 k         
#> 10 -0.409 -1.35  -1.53   -0.118 r         
#> # … with 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>       cd45     cd38   cd34    cd19 cluster_id
#>      <dbl>    <dbl>  <dbl>   <dbl> <chr>     
#>  1 -0.843   0.307    0.296  0.0982 l         
#>  2  1.43    0.00257  0.411  1.61   y         
#>  3  0.0585 -0.275    1.90  -0.819  r         
#>  4 -0.163   0.723   -1.51   0.977  q         
#>  5  0.972  -0.158   -0.243 -0.0326 b         
#>  6  0.0983  1.07     0.680 -0.580  g         
#>  7 -1.28    1.45    -0.746 -0.329  g         
#>  8 -1.51    0.136    0.685  0.314  o         
#>  9  1.24    0.546    1.35  -0.703  n         
#> 10  0.0855  0.403    0.342 -0.448  i         
#> # … with 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 90 × 5
#>       cd45    cd38    cd34     cd19 cluster_id
#>      <dbl>   <dbl>   <dbl>    <dbl> <chr>     
#>  1  1.09    1.12    0.0770  0.440   t         
#>  2  0.753   0.759   1.48    0.00887 h         
#>  3  0.335  -1.77    1.49   -0.226   u         
#>  4 -0.343  -0.482   0.0900 -0.264   t         
#>  5 -0.0527  0.0168 -0.380  -0.114   p         
#>  6  0.136  -1.32   -0.926  -1.28    r         
#>  7  0.447  -1.87    0.663   0.805   w         
#>  8 -1.88    2.87   -1.42   -0.0679  j         
#>  9 -0.350  -1.36    1.19   -1.45    q         
#> 10 -2.09   -1.82    0.817   0.106   o         
#> # … with 80 more rows