Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>      cd45   cd38    cd34    cd19 cluster_id
#>     <dbl>  <dbl>   <dbl>   <dbl> <chr>     
#>  1  0.388 -0.656  0.0169  2.22   l         
#>  2 -0.338 -0.203  0.311  -0.918  x         
#>  3 -0.668  1.42   0.663  -0.0426 n         
#>  4 -0.646  1.27   1.25   -1.39   b         
#>  5 -1.02  -0.942  1.73    0.199  r         
#>  6  1.39   0.103  0.630  -0.492  d         
#>  7 -0.803  0.217 -1.81   -0.404  h         
#>  8 -0.339 -1.35   1.01    1.22   r         
#>  9 -1.69   1.16  -2.18    0.152  e         
#> 10 -0.308 -0.304 -0.729  -0.194  q         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>       cd45     cd38   cd34    cd19 cluster_id
#>      <dbl>    <dbl>  <dbl>   <dbl> <chr>     
#>  1  0.554  -1.50    -3.22   0.551  z         
#>  2  0.0988 -1.64    -1.48  -1.42   f         
#>  3 -1.56    0.844   -0.890  0.0949 t         
#>  4 -1.40    0.00142 -0.277  1.58   t         
#>  5 -1.00   -1.95    -0.180  0.140  w         
#>  6  0.368   1.04    -0.811  1.51   f         
#>  7 -0.105  -0.0476   0.750 -0.0133 k         
#>  8  1.19    1.80    -0.989  0.135  t         
#>  9  1.24    1.26    -1.98   0.446  o         
#> 10 -0.824  -0.496   -0.775 -0.181  h         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 93 × 5
#>       cd45    cd38    cd34    cd19 cluster_id
#>      <dbl>   <dbl>   <dbl>   <dbl> <chr>     
#>  1 -0.338  -0.203   0.311  -0.918  x         
#>  2 -0.474  -1.24   -0.0408 -1.63   e         
#>  3  1.92   -0.278  -1.94    1.10   j         
#>  4  0.201  -0.887  -0.830  -0.175  c         
#>  5  1.33    1.39    0.283   0.0380 f         
#>  6 -1.36    1.40    1.22   -0.153  g         
#>  7  1.02    1.76    0.896  -1.99   h         
#>  8  0.0524  1.58   -0.660  -0.804  s         
#>  9  1.76   -0.0118  2.14    2.53   b         
#> 10  0.146   0.163  -0.148  -2.00   l         
#> # ℹ 83 more rows