Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>      cd45   cd38    cd34   cd19 cluster_id
#>     <dbl>  <dbl>   <dbl>  <dbl> <chr>     
#>  1 -0.593  1.31  -0.0865 -0.674 l         
#>  2 -0.594 -1.52  -0.232   0.829 s         
#>  3  0.531  0.622 -1.19   -0.779 q         
#>  4 -0.680 -0.376  0.647  -0.562 b         
#>  5 -1.31  -1.28  -0.589  -1.11  j         
#>  6 -0.556  0.640 -1.46   -1.20  h         
#>  7 -0.444 -0.324 -1.99    2.51  q         
#>  8 -0.926  0.305  0.116  -0.638 h         
#>  9  0.556  0.473 -0.690   0.265 c         
#> 10  0.252 -0.656 -2.02   -1.12  h         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>      cd45    cd38    cd34     cd19 cluster_id
#>     <dbl>   <dbl>   <dbl>    <dbl> <chr>     
#>  1  0.104 -0.874   0.817   0.841   k         
#>  2  0.367  1.31    0.100  -0.403   q         
#>  3 -0.652 -1.72    0.130   0.395   o         
#>  4  1.25  -0.204  -0.468  -1.38    p         
#>  5  0.207  1.13   -0.0486  0.0937  z         
#>  6 -1.17   2.35   -1.61   -2.10    f         
#>  7 -2.82  -0.164  -1.29   -0.675   j         
#>  8 -1.24   0.600  -0.244   0.00962 z         
#>  9  0.237  0.518  -1.10   -0.0750  g         
#> 10  0.632 -0.0485  0.816   1.77    g         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 122 × 5
#>      cd45    cd38   cd34   cd19 cluster_id
#>     <dbl>   <dbl>  <dbl>  <dbl> <chr>     
#>  1 -0.104  0.106   2.02   0.979 l         
#>  2 -0.926  0.305   0.116 -0.638 h         
#>  3  1.43  -0.423  -0.960  0.369 q         
#>  4  0.545  0.923   0.779  0.121 e         
#>  5  1.67  -1.39   -0.423  0.212 t         
#>  6 -1.24   0.664   0.432  0.812 v         
#>  7  1.73  -0.495   0.923  0.207 q         
#>  8  0.833 -0.0838  1.17   1.80  e         
#>  9  1.05   0.250   0.846  0.192 z         
#> 10  0.978 -0.375  -1.57   0.377 u         
#> # ℹ 112 more rows