Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>       cd45    cd38   cd34   cd19 cluster_id
#>      <dbl>   <dbl>  <dbl>  <dbl> <chr>     
#>  1 -0.0692  0.285   1.75  -0.343 a         
#>  2  0.302   2.12    0.830  0.868 s         
#>  3  0.564  -0.337  -0.804  1.05  i         
#>  4 -0.732   0.0620 -0.747  0.463 z         
#>  5 -0.280   1.33    0.405 -2.20  v         
#>  6  0.891   1.15   -0.613  0.358 r         
#>  7 -0.285   0.192  -0.505 -0.123 t         
#>  8 -1.28    0.837   0.399  0.329 t         
#>  9  0.532   1.61   -0.373  0.693 b         
#> 10 -0.391  -1.59    1.16  -0.363 y         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>        cd45     cd38     cd34     cd19 cluster_id
#>       <dbl>    <dbl>    <dbl>    <dbl> <chr>     
#>  1  0.416   -0.0549  -0.695    0.226   t         
#>  2 -0.811    0.802    0.674   -0.247   d         
#>  3  1.13    -0.218   -1.17    -0.354   l         
#>  4  0.514   -0.853    1.11    -0.959   j         
#>  5 -0.127   -0.115    0.840    0.696   p         
#>  6  0.00882  1.18     0.463    0.0690  l         
#>  7 -0.483    0.728   -0.210   -0.456   n         
#>  8  1.59    -1.13     0.00641  0.909   y         
#>  9  0.315    0.0125  -0.150    0.00140 s         
#> 10  0.147    0.00315 -1.83     0.743   j         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 95 × 5
#>       cd45   cd38     cd34    cd19 cluster_id
#>      <dbl>  <dbl>    <dbl>   <dbl> <chr>     
#>  1 -1.18   -0.355 -2.25    -0.514  g         
#>  2 -1.39   -0.266  0.671   -0.916  r         
#>  3 -0.0662 -0.249  0.407   -0.0341 s         
#>  4  0.324   0.379  0.697   -0.186  a         
#>  5 -2.17   -2.24  -0.895    0.114  a         
#>  6 -0.532   1.68   0.507    0.445  c         
#>  7 -0.848  -0.836  0.340   -0.572  h         
#>  8 -0.805  -1.24   0.00874  0.480  r         
#>  9 -0.298  -1.97   1.40    -0.383  z         
#> 10 -0.921   0.747  1.42    -1.29   w         
#> # ℹ 85 more rows