Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>       cd45    cd38   cd34   cd19 cluster_id
#>      <dbl>   <dbl>  <dbl>  <dbl> <chr>     
#>  1  1.04    0.0513 -0.134 -0.409 w         
#>  2  0.621  -0.243  -0.369  1.96  a         
#>  3 -1.84    0.123   0.953 -2.22  j         
#>  4  1.04    0.163   1.52   0.691 v         
#>  5  0.773  -0.327   0.935 -0.421 c         
#>  6  0.496   1.34    0.259  0.774 l         
#>  7  1.10   -0.394  -1.08  -0.660 i         
#>  8  0.933   0.689   0.209 -0.389 l         
#>  9 -0.0239 -0.362  -0.470  0.512 v         
#> 10  0.393  -0.912   1.17  -0.172 w         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>      cd45   cd38    cd34   cd19 cluster_id
#>     <dbl>  <dbl>   <dbl>  <dbl> <chr>     
#>  1 -1.92   2.01  -1.43   -0.488 p         
#>  2 -1.24   0.849  0.601  -0.535 z         
#>  3 -0.308  0.764  1.33    2.83  d         
#>  4  0.306 -0.295 -1.66   -0.782 f         
#>  5 -0.650  0.126  0.597  -1.55  x         
#>  6 -0.900 -2.25   0.708   0.417 p         
#>  7 -0.432  0.927  0.222   0.170 q         
#>  8  1.10   0.728 -0.0275  0.259 u         
#>  9  0.652  0.189  1.54    0.759 l         
#> 10 -0.464 -0.794  0.147  -1.86  a         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 117 × 5
#>       cd45   cd38   cd34   cd19 cluster_id
#>      <dbl>  <dbl>  <dbl>  <dbl> <chr>     
#>  1 -0.969  -0.969  0.700  1.52  j         
#>  2 -0.290   0.267  1.49   0.579 n         
#>  3  0.443   1.85   0.752  0.654 c         
#>  4  1.10   -0.394 -1.08  -0.660 i         
#>  5  0.0292 -0.718  1.01   0.133 i         
#>  6 -1.16   -0.555  0.428 -0.205 d         
#>  7 -1.80   -0.663  0.212 -2.07  o         
#>  8  1.58   -1.00  -0.296  2.02  w         
#>  9 -0.716   0.182  1.35  -0.234 s         
#> 10 -0.154   0.532  0.106 -0.553 y         
#> # ℹ 107 more rows