Downsample CyTOF data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>       cd45    cd38    cd34    cd19 cluster_id
#>      <dbl>   <dbl>   <dbl>   <dbl> <chr>     
#>  1  1.29   -0.0798 -0.0623 -0.251  q         
#>  2 -0.0979 -1.46    1.39    0.211  l         
#>  3 -0.499   0.313   0.688   0.857  t         
#>  4 -1.68    2.04    1.03   -0.672  o         
#>  5 -0.0532 -0.249  -0.542  -1.01   r         
#>  6  0.225   0.887  -0.0952  0.468  n         
#>  7  0.582  -0.917  -0.669   0.573  i         
#>  8  2.60    1.09    1.68   -0.344  c         
#>  9 -0.995   0.521   0.631  -0.0412 q         
#> 10 -1.29   -0.154  -0.857   1.39   q         
#> # … with 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>      cd45   cd38   cd34   cd19 cluster_id
#>     <dbl>  <dbl>  <dbl>  <dbl> <chr>     
#>  1  1.25  -1.36  -0.363 -0.825 k         
#>  2  0.737 -1.03   0.334 -0.762 o         
#>  3  0.365  1.03  -0.630 -1.36  n         
#>  4  1.33  -0.681 -0.675 -0.838 b         
#>  5  0.762 -0.398 -0.250 -1.42  a         
#>  6 -0.549 -1.01   0.258 -0.503 a         
#>  7  0.491  1.09  -0.444 -1.22  s         
#>  8  0.290  0.114  0.832 -1.11  m         
#>  9  0.944 -0.390 -0.738 -0.156 j         
#> 10 -0.371 -0.501 -0.867  1.71  k         
#> # … with 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 99 × 5
#>       cd45   cd38   cd34   cd19 cluster_id
#>      <dbl>  <dbl>  <dbl>  <dbl> <chr>     
#>  1  0.491   2.63  -1.54  -0.286 l         
#>  2 -0.0817 -0.250  0.651 -0.811 u         
#>  3 -0.484  -0.659  0.768  0.821 c         
#>  4 -1.04   -1.27  -0.352 -0.803 t         
#>  5 -0.699   2.01  -2.17   2.36  d         
#>  6  1.81    0.205  2.29   0.525 b         
#>  7 -0.620   2.88  -1.52   1.30  t         
#>  8 -0.966   0.850  0.645 -0.434 i         
#>  9 -1.89    0.667 -0.296 -0.913 m         
#> 10  1.48    0.322  0.464  0.437 r         
#> # … with 89 more rows