Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>      cd45    cd38    cd34   cd19 cluster_id
#>     <dbl>   <dbl>   <dbl>  <dbl> <chr>     
#>  1 -0.646  0.173   2.56   -0.515 c         
#>  2 -0.441 -0.622   0.630   0.810 y         
#>  3 -1.93   0.403  -0.381   0.497 o         
#>  4  0.262  1.88   -0.0172 -0.190 z         
#>  5 -1.17   1.34   -0.354  -0.869 n         
#>  6 -0.221  0.111   0.353   0.857 t         
#>  7 -1.46  -0.440   1.63   -0.239 i         
#>  8  1.63   1.11   -1.91    0.209 h         
#>  9 -0.983 -0.0646  2.42    1.86  t         
#> 10 -0.461 -0.759   0.753   0.918 x         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>      cd45    cd38   cd34    cd19 cluster_id
#>     <dbl>   <dbl>  <dbl>   <dbl> <chr>     
#>  1  0.251  0.763   1.50  -0.925  y         
#>  2 -0.901 -0.439  -1.09  -1.13   q         
#>  3  0.872  0.259  -1.63  -0.0570 z         
#>  4 -0.141  0.717   0.445  0.835  e         
#>  5 -1.04   0.0878  0.279  0.640  q         
#>  6  1.32   0.225  -0.245 -2.86   e         
#>  7 -0.779 -1.51    0.811 -0.826  v         
#>  8 -0.676 -1.05   -2.17  -1.49   d         
#>  9  0.316  0.404   1.10  -1.21   p         
#> 10  0.198 -0.799  -0.870  0.623  l         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 99 × 5
#>       cd45   cd38   cd34   cd19 cluster_id
#>      <dbl>  <dbl>  <dbl>  <dbl> <chr>     
#>  1  0.208  -0.517  0.601 -0.699 p         
#>  2  0.110  -0.871  0.252  0.127 o         
#>  3 -1.20   -0.254  1.01  -0.814 z         
#>  4  0.0467 -1.19  -0.946  0.949 j         
#>  5 -0.729   0.598 -0.147 -0.151 p         
#>  6 -0.221   1.12  -0.969  0.461 o         
#>  7 -1.45   -2.00   1.55   1.31  j         
#>  8 -1.36   -0.541  0.797  0.136 c         
#>  9 -1.44    0.817  0.295 -0.630 b         
#> 10 -1.38    1.22   0.603 -0.643 l         
#> # ℹ 89 more rows