Downsample high-dimensional cytometry data.

This function downsamples the number of cells in a `tof_tbl` using the one of three methods (randomly sampling a constant number of cells, randomly sampling a proportion of cells, or performing density-dependent downsampling per the algorithm in Qiu et al., (2011)).

Usage

tof_downsample(
  tof_tibble,
  group_cols = NULL,
  ...,
  method = c("constant", "prop", "density")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

...

Additional arguments to pass to the `tof_downsample_*` function family member corresponding to the chosen method.

method

A string indicating which downsampling method to use: "constant" (the default), "prop", or "density".

Value

A downsampled `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows in the result will depend on the chosen downsampling method.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 200 cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    num_cells = 200L,
    method = "constant"
)
#> # A tibble: 200 × 5
#>       cd45    cd38   cd34     cd19 cluster_id
#>      <dbl>   <dbl>  <dbl>    <dbl> <chr>     
#>  1 -0.976   1.90   -0.238  0.441   p         
#>  2 -0.195  -0.474  -2.22  -0.600   p         
#>  3 -0.212  -0.464  -1.20  -0.891   z         
#>  4  1.88    0.643   1.52  -0.869   b         
#>  5 -0.888  -0.231   0.321  0.0759  v         
#>  6  0.768  -0.0448 -1.85   0.525   x         
#>  7  0.0491 -1.82   -0.690 -0.507   b         
#>  8 -0.937   0.751  -1.21   0.252   l         
#>  9  0.473   0.261   0.298  1.00    f         
#> 10  0.457  -0.547   1.71   0.00101 x         
#> # ℹ 190 more rows

# sample 10% of all cells from the input data
tof_downsample(
    tof_tibble = sim_data,
    prop_cells = 0.1,
    method = "prop"
)
#> # A tibble: 100 × 5
#>       cd45   cd38   cd34   cd19 cluster_id
#>      <dbl>  <dbl>  <dbl>  <dbl> <chr>     
#>  1  0.0933 -0.218  0.840  0.164 k         
#>  2  0.0421  1.32  -1.24   0.163 e         
#>  3  1.27   -0.522 -0.135  1.48  u         
#>  4 -0.231  -0.223 -0.301  0.239 q         
#>  5 -0.721   0.722 -0.739  1.18  y         
#>  6 -0.0458  0.785  0.697  0.319 z         
#>  7  0.701  -0.605  1.84   0.861 r         
#>  8 -0.905  -0.783 -1.42  -0.385 s         
#>  9  0.867  -0.180  1.90   0.273 w         
#> 10  1.41    2.16   2.27   0.974 q         
#> # ℹ 90 more rows

# sample ~10% of cells from the input data using density dependence
tof_downsample(
    tof_tibble = sim_data,
    target_prop_cells = 0.1,
    method = "density"
)
#> # A tibble: 96 × 5
#>      cd45    cd38   cd34    cd19 cluster_id
#>     <dbl>   <dbl>  <dbl>   <dbl> <chr>     
#>  1  1.88   0.643   1.52  -0.869  b         
#>  2 -0.146  0.828   1.10  -0.0467 t         
#>  3 -1.18  -0.259  -0.383 -0.588  e         
#>  4 -1.86   0.0770  1.31   0.0992 l         
#>  5 -0.340  0.269  -0.969  1.63   m         
#>  6  0.979  0.370  -1.24   1.00   d         
#>  7 -0.241 -0.523   0.492  1.34   n         
#>  8 -1.90  -1.80   -1.81  -0.999  q         
#>  9 -0.448 -0.294   1.46   0.830  p         
#> 10 -0.191  0.737   1.10   0.0106 a         
#> # ℹ 86 more rows