Downsample CyTOF data by randomly selecting a constant number of cells per group.

This function downsamples the number of cells in a `tof_tbl` by randomly selecting `num_cells` cells from each unique combination of values in `group_cols`.

Usage

tof_downsample_constant(tof_tibble, group_cols = NULL, num_cells)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups from which `num_cells` will be downsampled. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

num_cells

An integer number of cells that should be sampled from each group defined by `group_cols`.

Value

A `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. Specifically, the number of rows will be `num_cells` multiplied by the number of unique combinations of the values in `group_cols`. If any group has fewer than `num_cells` number of cells, all cells from that group will be kept.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 500 cells from the input data
tof_downsample_constant(
    tof_tibble = sim_data,
    num_cells = 500L
)
#> # A tibble: 500 × 5
#>       cd45    cd38   cd34    cd19 cluster_id
#>      <dbl>   <dbl>  <dbl>   <dbl> <chr>     
#>  1 -0.0423 -1.44   -1.05  -0.567  w         
#>  2 -1.72    1.32    0.283 -0.448  b         
#>  3  0.741   0.0536  0.591  0.637  t         
#>  4 -1.93    0.937  -0.318  1.72   g         
#>  5 -1.83   -0.376  -1.32   1.17   s         
#>  6 -0.347  -0.0789 -0.897  0.755  i         
#>  7 -1.23   -2.37    0.134 -0.548  o         
#>  8  0.676   0.569  -0.803  0.0917 c         
#>  9 -0.435  -0.842  -0.132  2.36   k         
#> 10  0.857   0.476   0.730 -1.22   h         
#> # … with 490 more rows

# sample 20 cells per cluster from the input data
tof_downsample_constant(
    tof_tibble = sim_data,
    group_cols = cluster_id,
    num_cells = 20L
)
#> # A tibble: 520 × 5
#>       cd45    cd38    cd34    cd19 cluster_id
#>      <dbl>   <dbl>   <dbl>   <dbl> <chr>     
#>  1 -0.0423 -1.44   -1.05   -0.567  w         
#>  2 -1.42   -0.903  -1.14    0.520  c         
#>  3 -0.598   1.81    1.71    1.13   d         
#>  4 -1.72    1.32    0.283  -0.448  b         
#>  5 -1.83   -0.376  -1.32    1.17   s         
#>  6 -0.347  -0.0789 -0.897   0.755  i         
#>  7 -0.359  -0.0267 -1.41   -2.02   m         
#>  8  0.676   0.569  -0.803   0.0917 c         
#>  9  0.0245 -0.561  -0.0468 -0.979  t         
#> 10  0.857   0.476   0.730  -1.22   h         
#> # … with 510 more rows