Downsample high-dimensional cytometry data by randomly selecting a proportion of the cells in each group.

This function downsamples the number of cells in a `tof_tbl` using the density-dependent downsampling algorithm described in Qiu et al., (2011).

Usage

tof_downsample_density(
  tof_tibble,
  group_cols = NULL,
  density_cols = where(tof_is_numeric),
  target_num_cells,
  target_prop_cells,
  target_percentile = 0.03,
  outlier_percentile = 0.01,
  distance_function = c("euclidean", "cosine", "l2", "ip"),
  density_estimation_method = c("mean_distance", "sum_distance", "spade"),
  ...
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups within which the downsampling will be performed. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

density_cols

Unquoted names of the columns in `tof_tibble` to use in the density estimation for each cell. Defaults to all numeric columns in `tof_tibble`.

target_num_cells

An approximate constant number of cells (between 0 and 1) that should be sampled from each group defined by `group_cols`. Slightly more or fewer cells may be returned due to how the density calculation is performed.

target_prop_cells

An approximate proportion of cells (between 0 and 1) that should be sampled from each group defined by `group_cols`. Slightly more or fewer cells may be returned due to how the density calculation is performed. Ignored if `target_num_cells` is specified.

target_percentile

The local density percentile (i.e. a value between 0 and 1) to which the downsampling procedure should adjust all cells. In short, the algorithm will continue to remove cells from the input `tof_tibble` until the local densities of all remaining cells is equal to `target_percentile`. Lower values will result in more cells being removed. See Qiu et al., (2011) for details. Defaults to 0.1 (the 10th percentile of local densities). Ignored if either `target_num_cells` or `target_prop_cells` are specified.

outlier_percentile

The local density percentile (i.e. a value between 0 and 1) below which cells should be considered outliers (and discarded). Cells with a local density below `outlier_percentile` will never be selected during the downsampling procedure. Defaults to 0.01 (cells below the 1st local density percentile will be removed).

distance_function

A string indicating which distance function to use for the cell-to-cell distance calculations. Options include "euclidean" (the default) and "cosine" distances.

density_estimation_method

A string indicating which algorithm should be used to calculate the local density estimate for each cell. Options include k-nearest neighbor density estimation using the mean distance to a cell's k-nearest neighbors ("mean_distance"; the default), k-nearest neighbor density estimation using the summed distance to a cell's k nearest neighbors ("sum_distance") and counting the number of neighboring cells within a spherical radius around each cell as described in Qiu et al., 2011 ("spade"). While "spade" often produces the best results, it is slower than knn-density estimation methods.

...

Optional additional arguments to pass to tof_knn_density or tof_spade_density.

Value

A `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. The number of rows will depend on the chosen value of `target_percentile`, with fewer cells selected with lower values of `target_percentile`.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000)
    )

tof_downsample_density(
    tof_tibble = sim_data,
    density_cols = c(cd45, cd34, cd38),
    target_prop_cells = 0.5,
    density_estimation_method = "spade"
)
#> # A tibble: 513 × 4
#>      cd45    cd38   cd34   cd19
#>     <dbl>   <dbl>  <dbl>  <dbl>
#>  1 -0.285  0.0704 -0.150 -0.590
#>  2 -1.21  -0.631   2.19  -1.39 
#>  3 -1.06  -1.25    0.112 -0.366
#>  4  0.152  1.29    0.562  0.945
#>  5 -1.32   0.478  -0.726 -0.916
#>  6  1.46   1.46    0.696 -2.19 
#>  7  0.193 -0.270  -1.52  -1.42 
#>  8 -0.204 -0.583   0.124 -0.474
#>  9 -1.57   0.320  -1.47   1.47 
#> 10 -1.38  -0.438  -0.630 -1.26 
#> # ℹ 503 more rows

tof_downsample_density(
    tof_tibble = sim_data,
    density_cols = c(cd45, cd34, cd38),
    target_num_cells = 200L,
    density_estimation_method = "spade"
)
#> # A tibble: 187 × 4
#>      cd45    cd38    cd34    cd19
#>     <dbl>   <dbl>   <dbl>   <dbl>
#>  1  0.488  0.200  -1.26   -0.0137
#>  2 -1.38  -0.438  -0.630  -1.26  
#>  3 -1.41   1.26    0.0965  0.708 
#>  4 -0.717 -0.0211 -0.859   0.0695
#>  5 -0.157  1.25   -0.411   0.267 
#>  6 -2.46  -0.0429 -1.07    1.69  
#>  7  0.404  0.0739  2.00    0.327 
#>  8  0.285  0.981   0.804   1.62  
#>  9 -0.302 -1.02    1.56    0.698 
#> 10  0.320 -1.40   -1.50    0.925 
#> # ℹ 177 more rows

tof_downsample_density(
    tof_tibble = sim_data,
    density_cols = c(cd45, cd34, cd38),
    target_num_cells = 200L,
    density_estimation_method = "mean_distance"
)
#> # A tibble: 208 × 4
#>      cd45    cd38    cd34    cd19
#>     <dbl>   <dbl>   <dbl>   <dbl>
#>  1  1.46   1.46    0.696  -2.19  
#>  2 -0.672 -0.516  -0.196   1.04  
#>  3  0.423  2.45   -1.62    0.936 
#>  4  1.86  -2.28   -0.766   0.643 
#>  5 -1.69  -1.10   -0.951  -1.22  
#>  6 -0.729  1.84    0.143  -0.297 
#>  7 -1.14  -0.379  -1.29   -0.0373
#>  8  0.687 -0.0468 -0.469   0.280 
#>  9 -0.346  1.09   -0.604  -1.58  
#> 10  1.07  -2.00    0.0471 -0.819 
#> # ℹ 198 more rows