Downsample high-dimensional cytometry data by randomly selecting a proportion of the cells in each group.

This function downsamples the number of cells in a `tof_tbl` by randomly selecting a `prop_cells` proportion of the total number of cells with each unique combination of values in `group_cols`.

Usage

tof_downsample_prop(tof_tibble, group_cols = NULL, prop_cells)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

group_cols

Unquoted names of the columns in `tof_tibble` that should be used to define groups from which `prop_cells` will be downsampled. Supports tidyselect helpers. Defaults to `NULL` (no grouping).

prop_cells

A proportion of cells (between 0 and 1) that should be sampled from each group defined by `group_cols`.

Value

A `tof_tbl` with the same number of columns as the input `tof_tibble`, but fewer rows. Specifically, the number of rows should be `prop_cells` times the number of rows in the input `tof_tibble`.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

# sample 10% of all cells from the input data
tof_downsample_prop(
    tof_tibble = sim_data,
    prop_cells = 0.1
)
#> # A tibble: 100 × 5
#>       cd45    cd38   cd34    cd19 cluster_id
#>      <dbl>   <dbl>  <dbl>   <dbl> <chr>     
#>  1 -0.885   1.27    0.269  0.280  s         
#>  2  0.804   0.785  -0.735  0.313  d         
#>  3 -0.837   1.06   -0.372 -0.378  c         
#>  4 -0.0447 -0.890   0.236  0.692  t         
#>  5 -0.623   0.0114 -0.784  1.64   w         
#>  6 -0.244   0.117  -0.156 -0.268  m         
#>  7  1.43    0.159   0.708  0.136  i         
#>  8  0.693   0.312   2.10   0.252  u         
#>  9  0.0375  0.0720  1.26   1.10   p         
#> 10 -0.0276 -2.79    1.75   0.0810 f         
#> # ℹ 90 more rows

# sample 10% of all cells from each cluster in the input data
tof_downsample_prop(
    tof_tibble = sim_data,
    group_cols = cluster_id,
    prop_cells = 0.1
)
#> # A tibble: 89 × 5
#>       cd45     cd38    cd34    cd19 cluster_id
#>      <dbl>    <dbl>   <dbl>   <dbl> <chr>     
#>  1 -1.71   -0.432   -1.34   -1.38   a         
#>  2 -0.351  -0.832    1.61    1.31   a         
#>  3 -0.845  -1.09    -0.191  -1.36   a         
#>  4 -0.663  -0.784   -0.831   1.20   b         
#>  5  0.700  -0.00106  0.583   0.793  b         
#>  6  2.51   -1.03     1.02    0.242  b         
#>  7 -1.10   -1.02    -0.984   0.0933 c         
#>  8 -0.736   0.00822 -0.0429  0.492  c         
#>  9 -0.0943  0.592    0.293   1.07   c         
#> 10  0.323   1.59     0.716  -0.671  d         
#> # ℹ 79 more rows