Estimate the local densities for all cells in a CyTOF dataset.

This function is a wrapper around tidytof's tof_*_density() function family. It performs local density estimation on CyTOF data using a user-specified method (of 3 choices) and each method's corresponding input parameters.

Usage

tof_estimate_density(
  tof_tibble,
  distance_cols = where(tof_is_numeric),
  distance_function = c("euclidean", "cosine"),
  normalize = TRUE,
  ...,
  augment = TRUE,
  method = c("mean_distance", "sum_distance", "spade")
)

Arguments

tof_tibble

A `tof_tbl` or a `tibble`.

distance_cols

Unquoted names of the columns in `tof_tibble` to use in calculating cell-to-cell distances during the local density estimation for each cell. Defaults to all numeric columns in `tof_tibble`.

distance_function

A string indicating which distance function to use for calculating cell-to-cell distances during local density estimation. Options include "euclidean" (the default) and "cosine".

normalize

A boolean value indicating if the vector of local density estimates should be normalized to values between 0 and 1. Defaults to TRUE.

...

Additional arguments to pass to the `tof_*_density()` function family member corresponding to the chosen `method`.

augment

A boolean value indicating if the output should column-bind the local density estimates of each cell as a new column in `tof_tibble` (TRUE; the default) or if a single-column tibble including only the local density estimates should be returned (FALSE).

method

A string indicating which local density estimation method should be used. Valid values include "mean_distance", "sum_distance", and "spade".

Value

A `tof_tbl` or `tibble` If augment = FALSE, it will have a single column encoding the local density estimates for each cell in `tof_tibble`. If augment = TRUE, it will have ncol(tof_tibble) + 1 columns: each of the (unaltered) columns in `tof_tibble` plus an additional column encoding the local density estimates.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000)
    )

# perform the density estimation
tof_estimate_density(tof_tibble = sim_data, method = "spade")
#> # A tibble: 1,000 × 5
#>      cd45    cd38   cd34   cd19 .spade_density
#>     <dbl>   <dbl>  <dbl>  <dbl>          <dbl>
#>  1  1.73  -0.580  -0.288 -0.638              1
#>  2  0.985 -0.0288  0.146 -0.574              1
#>  3 -0.794 -1.14   -1.87   0.649              1
#>  4  0.595 -2.40    0.992  0.104              1
#>  5 -0.794 -0.208   0.379  1.08               1
#>  6 -2.32   0.270  -0.717 -0.600              1
#>  7  0.503 -0.871  -0.544 -1.05               1
#>  8 -0.530  1.03   -0.948 -1.20               1
#>  9  0.837  0.601  -0.657 -0.548              1
#> 10  0.613  0.0395  0.377 -0.160              1
#> # … with 990 more rows

# perform the density estimation with a smaller search radius around
# each cell
tof_estimate_density(
    tof_tibble = sim_data,
    alpha_multiplier = 2,
    method = "spade"
)
#> # A tibble: 1,000 × 5
#>      cd45    cd38   cd34   cd19 .spade_density
#>     <dbl>   <dbl>  <dbl>  <dbl>          <dbl>
#>  1  1.73  -0.580  -0.288 -0.638            0.5
#>  2  0.985 -0.0288  0.146 -0.574            1  
#>  3 -0.794 -1.14   -1.87   0.649            0.3
#>  4  0.595 -2.40    0.992  0.104            0  
#>  5 -0.794 -0.208   0.379  1.08             1  
#>  6 -2.32   0.270  -0.717 -0.600            0.3
#>  7  0.503 -0.871  -0.544 -1.05             1  
#>  8 -0.530  1.03   -0.948 -1.20             0.8
#>  9  0.837  0.601  -0.657 -0.548            1  
#> 10  0.613  0.0395  0.377 -0.160            1  
#> # … with 990 more rows