Metacluster clustered CyTOF data.

This function is a wrapper around tidytof's tof_metacluster_* function family. It performs metaclustering on CyTOF data using a user-specified method (of 5 choices) and each method's corresponding input parameters.

Usage

tof_metacluster(
  tof_tibble,
  cluster_col,
  metacluster_cols = where(tof_is_numeric),
  central_tendency_function = stats::median,
  ...,
  augment = TRUE,
  method = c("consensus", "hierarchical", "kmeans", "phenograph", "flowsom")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

cluster_col

An unquoted column name indicating which column in `tof_tibble` stores the cluster ids for the cluster to which each cell belongs. Cluster labels can be produced via any method the user chooses - including manual gating, any of the functions in the `tof_cluster_*` function family, or any other method.

metacluster_cols

Unquoted column names indicating which columns in `tof_tibble` to use in computing the metaclusters. Defaults to all numeric columns in `tof_tibble`. Supports tidyselect helpers.

central_tendency_function

The function that should be used to calculate the measurement of central tendency for each cluster before metaclustering. This function will be used to compute a summary statistic for each input cluster in `cluster_col` across all columns specified by `metacluster_cols`, and the resulting vector (one for each cluster) will be used as the input for metaclustering. Defaults to median.

...

Additional arguments to pass to the `tof_metacluster_*` function family member corresponding to the chosen `method`.

augment

A boolean value indicating if the output should column-bind the metacluster ids of each cell as a new column in `tof_tibble` (TRUE; the default) or if a single-column tibble including only the metacluster ids should be returned (FALSE).

method

A string indicating which clustering method should be used. Valid values include "consensus", "hierarchical", "kmeans", "phenograph", and "flowsom".

Value

A `tof_tbl` or `tibble` If augment = FALSE, it will have a single column encoding the metacluster ids for each cell in `tof_tibble`. If augment = TRUE, it will have ncol(tof_tibble) + 1 columns: each of the (unaltered) columns in `tof_tibble` plus an additional column encoding the metacluster ids.

See also

Examples

sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 1000),
        cd38 = rnorm(n = 1000),
        cd34 = rnorm(n = 1000),
        cd19 = rnorm(n = 1000),
        cluster_id = sample(letters, size = 1000, replace = TRUE)
    )

tof_metacluster(
    tof_tibble = sim_data,
    cluster_col = cluster_id,
    clustering_algorithm = "consensus",
    method = "flowsom"
)
#> # A tibble: 1,000 × 6
#>      cd45    cd38    cd34   cd19 cluster_id .flowsom_metacluster
#>     <dbl>   <dbl>   <dbl>  <dbl> <chr>      <chr>               
#>  1  0.764 -1.01   -2.27   -0.449 i          7                   
#>  2  0.398  1.64   -2.25    0.791 i          7                   
#>  3  0.276  1.02    0.766   0.911 f          4                   
#>  4 -2.01  -1.61   -0.0797 -0.873 z          2                   
#>  5 -0.908  0.237  -0.987   1.76  l          4                   
#>  6 -0.564  2.38   -1.37    0.249 m          1                   
#>  7  1.58   0.453   0.252   1.24  v          4                   
#>  8  0.583 -1.98    0.318   0.787 w          3                   
#>  9 -0.529  0.0542  0.855  -1.76  h          6                   
#> 10 -0.620 -1.91    0.653  -0.796 w          3                   
#> # … with 990 more rows

tof_metacluster(
    tof_tibble = sim_data,
    cluster_col = cluster_id,
    method = "phenograph"
)
#> # A tibble: 1,000 × 6
#>      cd45    cd38    cd34   cd19 cluster_id .phenograph_metacluster
#>     <dbl>   <dbl>   <dbl>  <dbl> <chr>      <chr>                  
#>  1  0.764 -1.01   -2.27   -0.449 i          2                      
#>  2  0.398  1.64   -2.25    0.791 i          2                      
#>  3  0.276  1.02    0.766   0.911 f          3                      
#>  4 -2.01  -1.61   -0.0797 -0.873 z          1                      
#>  5 -0.908  0.237  -0.987   1.76  l          3                      
#>  6 -0.564  2.38   -1.37    0.249 m          5                      
#>  7  1.58   0.453   0.252   1.24  v          3                      
#>  8  0.583 -1.98    0.318   0.787 w          2                      
#>  9 -0.529  0.0542  0.855  -1.76  h          1                      
#> 10 -0.620 -1.91    0.653  -0.796 w          2                      
#> # … with 990 more rows