Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>       cd45   cd38    cd34     cd19    .pc1   .pc2   .pc3   .pc4
#>      <dbl>  <dbl>   <dbl>    <dbl>   <dbl>  <dbl>  <dbl>  <dbl>
#>  1 -1.06   -1.25  -0.230  -0.613    0.353   1.38  -1.09   0.172
#>  2 -0.678   2.03  -1.04   -0.00202 -0.0409 -0.880  0.603  2.30 
#>  3  0.700   0.419  0.0472 -1.08     0.959  -0.899 -0.493  0.130
#>  4 -0.0907  0.177  0.824  -0.898    0.0328 -0.638 -1.05  -0.109
#>  5 -0.627   0.934  0.143  -0.255   -0.355  -0.571 -0.312  0.936
#>  6 -2.48   -0.180  1.69    0.128   -2.21    0.569 -1.65   0.230
#>  7  1.54    0.512  0.599  -1.20     1.09   -1.57  -0.462 -0.572
#>  8  0.207   0.940  1.30    0.980   -1.34   -1.04   0.528 -0.390
#>  9  0.526  -1.97  -0.0926  0.0210   0.882   1.42  -0.161 -1.25 
#> 10  0.175  -0.624 -1.38   -1.71     2.17    0.532 -0.847  0.774
#> # … with 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>       cd45   cd38    cd34     cd19 .tsne_1 .tsne_2
#>      <dbl>  <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1 -1.06   -1.25  -0.230  -0.613     2.80    1.05 
#>  2 -0.678   2.03  -1.04   -0.00202   2.94    4.10 
#>  3  0.700   0.419  0.0472 -1.08      0.140  -2.06 
#>  4 -0.0907  0.177  0.824  -0.898    -0.911  -0.693
#>  5 -0.627   0.934  0.143  -0.255    -2.08   -0.137
#>  6 -2.48   -0.180  1.69    0.128    -3.25    2.57 
#>  7  1.54    0.512  0.599  -1.20     -1.00   -2.34 
#>  8  0.207   0.940  1.30    0.980    -3.53   -0.183
#>  9  0.526  -1.97  -0.0926  0.0210    1.34    0.156
#> 10  0.175  -0.624 -1.38   -1.71      3.15   -1.86 
#> # … with 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> # A tibble: 100 × 6
#>       cd45   cd38    cd34     cd19  .umap1 .umap2
#>      <dbl>  <dbl>   <dbl>    <dbl>   <dbl>  <dbl>
#>  1 -1.06   -1.25  -0.230  -0.613   -2.55    2.51 
#>  2 -0.678   2.03  -1.04   -0.00202  3.80   -2.88 
#>  3  0.700   0.419  0.0472 -1.08    -0.175  -0.457
#>  4 -0.0907  0.177  0.824  -0.898   -0.826   2.42 
#>  5 -0.627   0.934  0.143  -0.255    3.46   -2.50 
#>  6 -2.48   -0.180  1.69    0.128    2.05   -4.13 
#>  7  1.54    0.512  0.599  -1.20    -0.693  -0.405
#>  8  0.207   0.940  1.30    0.980    0.177  -2.82 
#>  9  0.526  -1.97  -0.0926  0.0210  -2.04    1.78 
#> 10  0.175  -0.624 -1.38   -1.71    -0.0169  0.908
#> # … with 90 more rows