Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>      cd45   cd38   cd34    cd19   .pc1   .pc2   .pc3    .pc4
#>     <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1  0.155  1.57   1.37   0.358  -0.401  1.76  -1.02   0.792 
#>  2  0.441  1.14  -0.162 -0.959   0.761  0.204 -1.21  -0.0295
#>  3 -1.70   0.350 -0.274 -0.265  -0.793  0.240 -0.314 -1.50  
#>  4  0.223  0.301  0.801 -1.09   -0.299 -0.148 -0.898  0.670 
#>  5 -0.150 -1.15  -0.675  0.712   0.213 -0.186  1.50  -0.523 
#>  6  0.754  0.719  0.127 -0.755   0.666  0.143 -0.828  0.460 
#>  7 -1.32  -1.45   0.443  0.530  -1.47  -0.164  1.32  -0.454 
#>  8  0.789  0.475  0.787  0.0188  0.174  0.683 -0.311  0.983 
#>  9 -0.776 -0.897  0.789 -0.274  -1.24  -0.300  0.395  0.125 
#> 10  0.503  0.484 -0.660 -0.746   1.01  -0.186 -0.514 -0.261 
#> # … with 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>      cd45   cd38   cd34    cd19 .tsne_1 .tsne_2
#>     <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>
#>  1  0.155  1.57   1.37   0.358     4.34  -0.464
#>  2  0.441  1.14  -0.162 -0.959     3.10  -1.84 
#>  3 -1.70   0.350 -0.274 -0.265    -2.64   1.60 
#>  4  0.223  0.301  0.801 -1.09      2.68   0.518
#>  5 -0.150 -1.15  -0.675  0.712    -2.56  -1.29 
#>  6  0.754  0.719  0.127 -0.755     2.65  -1.46 
#>  7 -1.32  -1.45   0.443  0.530    -3.67   2.40 
#>  8  0.789  0.475  0.787  0.0188    2.58  -0.472
#>  9 -0.776 -0.897  0.789 -0.274    -1.20   2.29 
#> 10  0.503  0.484 -0.660 -0.746     1.81  -2.35 
#> # … with 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> # A tibble: 100 × 6
#>      cd45   cd38   cd34    cd19 .umap1 .umap2
#>     <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#>  1  0.155  1.57   1.37   0.358  -3.36  -0.716
#>  2  0.441  1.14  -0.162 -0.959  -2.28  -1.34 
#>  3 -1.70   0.350 -0.274 -0.265   0.927  2.25 
#>  4  0.223  0.301  0.801 -1.09   -2.04  -2.46 
#>  5 -0.150 -1.15  -0.675  0.712   1.45   0.344
#>  6  0.754  0.719  0.127 -0.755  -1.54  -2.43 
#>  7 -1.32  -1.45   0.443  0.530   2.64  -0.445
#>  8  0.789  0.475  0.787  0.0188 -0.792 -2.97 
#>  9 -0.776 -0.897  0.789 -0.274   2.79   0.716
#> 10  0.503  0.484 -0.660 -0.746  -1.96  -0.694
#> # … with 90 more rows