Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>      cd45   cd38   cd34    cd19   .pc1    .pc2     .pc3   .pc4
#>     <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>    <dbl>  <dbl>
#>  1 -0.971  1.69   0.584 -0.305  -1.48   0.783   0.523   -1.21 
#>  2 -1.06  -0.420 -1.15   1.42   -0.271 -2.17    0.0324  -0.738
#>  3  0.114 -0.447  1.69   1.93    1.86  -0.0560 -0.00542 -1.87 
#>  4  0.827 -0.849  0.356 -0.430   0.639  0.498  -0.285    0.832
#>  5 -0.225  1.13  -0.796  1.36   -0.338 -1.20    1.46    -0.902
#>  6 -0.910  1.92  -0.884 -0.235  -2.07  -0.259   1.16    -0.514
#>  7  0.463  0.464  1.17  -0.0357  0.354  1.12    0.344   -0.467
#>  8  1.46  -2.30   1.45  -0.312   2.17   1.01   -1.25     1.00 
#>  9 -0.423 -1.49  -2.07   0.628  -0.194 -2.49   -0.374    0.950
#> 10 -1.93  -0.713  1.55  -1.29   -1.19   1.14   -2.30    -0.883
#> # … with 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>      cd45   cd38   cd34    cd19 .tsne_1 .tsne_2
#>     <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>
#>  1 -0.971  1.69   0.584 -0.305   -4.18   -0.222
#>  2 -1.06  -0.420 -1.15   1.42    -0.888  -3.79 
#>  3  0.114 -0.447  1.69   1.93     3.30   -0.910
#>  4  0.827 -0.849  0.356 -0.430    1.98    1.71 
#>  5 -0.225  1.13  -0.796  1.36    -3.01   -2.22 
#>  6 -0.910  1.92  -0.884 -0.235   -4.25   -1.36 
#>  7  0.463  0.464  1.17  -0.0357  -1.24    1.15 
#>  8  1.46  -2.30   1.45  -0.312    4.06    0.927
#>  9 -0.423 -1.49  -2.07   0.628    0.111  -4.68 
#> 10 -1.93  -0.713  1.55  -1.29     3.14   -3.36 
#> # … with 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> # A tibble: 100 × 6
#>      cd45   cd38   cd34    cd19 .umap1  .umap2
#>     <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>
#>  1 -0.971  1.69   0.584 -0.305   1.40   3.54  
#>  2 -1.06  -0.420 -1.15   1.42   -3.67  -1.15  
#>  3  0.114 -0.447  1.69   1.93    0.876 -0.918 
#>  4  0.827 -0.849  0.356 -0.430  -0.937 -1.34  
#>  5 -0.225  1.13  -0.796  1.36    2.46   0.0745
#>  6 -0.910  1.92  -0.884 -0.235   0.452  3.40  
#>  7  0.463  0.464  1.17  -0.0357  2.11  -1.09  
#>  8  1.46  -2.30   1.45  -0.312  -1.45  -1.49  
#>  9 -0.423 -1.49  -2.07   0.628  -4.49  -0.484 
#> 10 -1.93  -0.713  1.55  -1.29    0.341  0.782 
#> # … with 90 more rows