Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>       cd45   cd38   cd34   cd19    .pc1    .pc2    .pc3     .pc4
#>      <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>    <dbl>
#>  1  0.576  -0.753  0.234  0.778 -0.724  -0.467   0.463  -0.923  
#>  2 -0.598   0.285  2.03   1.48  -2.01   -1.91    0.158   0.840  
#>  3 -1.15   -1.19  -1.34   1.05  -1.19    1.07   -1.43   -1.25   
#>  4 -0.304   0.553 -0.666  0.322  0.224  -0.0120 -0.882  -0.00612
#>  5 -0.0338  0.226 -0.560 -0.731  0.585   0.674  -0.0154  0.258  
#>  6  0.774   0.867 -1.32  -0.504  1.68    0.278  -0.402  -0.306  
#>  7 -0.880  -0.122 -0.379 -0.271 -0.361   0.789  -0.461   0.359  
#>  8 -0.331   2.25   0.150  0.906  0.495  -1.65   -1.37    1.09   
#>  9 -0.198   0.877  0.670 -0.274  0.0215 -0.519   0.164   1.12   
#> 10 -0.756  -0.173  1.29   1.31  -1.89   -1.12   -0.123   0.356  
#> # ℹ 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>       cd45   cd38   cd34   cd19 .tsne1  .tsne2
#>      <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1  0.576  -0.753  0.234  0.778  0.327  3.48  
#>  2 -0.598   0.285  2.03   1.48  -2.39   3.02  
#>  3 -1.15   -1.19  -1.34   1.05   3.46   2.95  
#>  4 -0.304   0.553 -0.666  0.322 -1.22  -0.0279
#>  5 -0.0338  0.226 -0.560 -0.731  1.66  -1.54  
#>  6  0.774   0.867 -1.32  -0.504  2.33  -3.41  
#>  7 -0.880  -0.122 -0.379 -0.271  2.65   1.24  
#>  8 -0.331   2.25   0.150  0.906 -3.47  -0.389 
#>  9 -0.198   0.877  0.670 -0.274 -1.79  -0.249 
#> 10 -0.756  -0.173  1.29   1.31  -1.94   2.75  
#> # ℹ 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> # A tibble: 100 × 6
#>       cd45   cd38   cd34   cd19 .umap1 .umap2
#>      <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1  0.576  -0.753  0.234  0.778  1.96  -2.53 
#>  2 -0.598   0.285  2.03   1.48   2.72  -0.958
#>  3 -1.15   -1.19  -1.34   1.05   1.48   0.188
#>  4 -0.304   0.553 -0.666  0.322  2.10   1.51 
#>  5 -0.0338  0.226 -0.560 -0.731 -2.21  -0.740
#>  6  0.774   0.867 -1.32  -0.504 -3.95  -1.12 
#>  7 -0.880  -0.122 -0.379 -0.271  2.13   0.700
#>  8 -0.331   2.25   0.150  0.906  0.542  3.67 
#>  9 -0.198   0.877  0.670 -0.274 -0.461  2.06 
#> 10 -0.756  -0.173  1.29   1.31   2.55  -1.05 
#> # ℹ 90 more rows