Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>      cd45    cd38    cd34    cd19   .pc1   .pc2   .pc3   .pc4
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1  1.39  -0.189  -0.202  -0.956  -0.752  0.370 -1.34  -0.395
#>  2  1.39  -1.25   -2.00   -0.963   0.656 -0.729 -2.61  -0.274
#>  3 -0.688 -0.896   0.0756  0.610   0.743 -0.189  0.720 -0.455
#>  4 -1.18  -0.0477 -1.78   -0.249   1.17  -0.839 -0.456  1.45 
#>  5  0.925  0.198  -0.208  -0.0243 -0.801 -0.288 -0.504 -0.217
#>  6  0.388 -0.689   0.252   1.34   -0.118 -0.791  0.672 -1.17 
#>  7 -0.207 -1.14    1.05    0.931   0.259  0.189  1.18  -1.39 
#>  8  0.947  2.46    0.174   1.87   -2.63  -1.68   1.15   0.439
#>  9 -0.258 -2.54   -0.769  -1.52    2.07   0.916 -1.49  -0.655
#> 10  1.77   0.659   0.726  -0.134  -1.95   0.245 -0.390 -0.724
#> # ℹ 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>      cd45    cd38    cd34    cd19 .tsne1 .tsne2
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>
#>  1  1.39  -0.189  -0.202  -0.956   1.44  -0.343
#>  2  1.39  -1.25   -2.00   -0.963  -2.51   2.47 
#>  3 -0.688 -0.896   0.0756  0.610  -2.30  -1.68 
#>  4 -1.18  -0.0477 -1.78   -0.249  -2.58   0.667
#>  5  0.925  0.198  -0.208  -0.0243  1.65   1.49 
#>  6  0.388 -0.689   0.252   1.34   -2.63  -3.18 
#>  7 -0.207 -1.14    1.05    0.931  -1.81  -3.71 
#>  8  0.947  2.46    0.174   1.87    2.89   4.60 
#>  9 -0.258 -2.54   -0.769  -1.52    0.474 -3.26 
#> 10  1.77   0.659   0.726  -0.134   3.42   2.05 
#> # ℹ 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> # A tibble: 100 × 6
#>      cd45    cd38    cd34    cd19 .umap1 .umap2
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>
#>  1  1.39  -0.189  -0.202  -0.956  -3.80  -2.20 
#>  2  1.39  -1.25   -2.00   -0.963   1.54  -1.89 
#>  3 -0.688 -0.896   0.0756  0.610   1.52   0.938
#>  4 -1.18  -0.0477 -1.78   -0.249   0.418 -0.149
#>  5  0.925  0.198  -0.208  -0.0243 -3.80  -1.13 
#>  6  0.388 -0.689   0.252   1.34    3.91  -2.45 
#>  7 -0.207 -1.14    1.05    0.931   4.12  -3.12 
#>  8  0.947  2.46    0.174   1.87   -2.15   1.50 
#>  9 -0.258 -2.54   -0.769  -1.52   -1.83  -2.71 
#> 10  1.77   0.659   0.726  -0.134  -4.28   0.887
#> # ℹ 90 more rows