Apply dimensionality reduction to a CyTOF dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on CyTOF data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>      cd45   cd38   cd34   cd19    .pc1   .pc2    .pc3   .pc4
#>     <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>  <dbl>
#>  1 -0.195 -0.482  0.943  0.475  0.130   1.01   0.176   0.553
#>  2  0.355  0.197 -0.411  0.426  0.615  -0.564  0.0148  0.116
#>  3  1.67   0.477  1.14   0.623  0.217  -0.151  1.76    1.23 
#>  4 -0.212 -2.68   0.392  0.421  1.35    2.27   0.176  -1.01 
#>  5 -0.221  0.799  0.184  0.581  0.0182 -0.412 -0.329   0.923
#>  6 -1.03   0.553 -0.847  0.410  0.293  -0.666 -1.42    0.119
#>  7 -0.447 -1.27   0.529 -0.429 -0.127   1.31   0.0252 -0.802
#>  8 -0.167 -0.128 -0.837  1.28   1.55   -0.349 -0.740   0.388
#>  9  1.39   1.35  -0.130 -1.55  -1.41   -1.76   1.27   -0.677
#> 10  0.876  1.16  -0.158  1.50   1.09   -1.17   0.282   1.60 
#> # … with 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>      cd45   cd38   cd34   cd19 .tsne_1 .tsne_2
#>     <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>
#>  1 -0.195 -0.482  0.943  0.475  -0.665   0.539
#>  2  0.355  0.197 -0.411  0.426   1.00   -0.523
#>  3  1.67   0.477  1.14   0.623  -3.60    2.81 
#>  4 -0.212 -2.68   0.392  0.421  -4.12   -2.88 
#>  5 -0.221  0.799  0.184  0.581   1.74    0.866
#>  6 -1.03   0.553 -0.847  0.410   2.94   -1.38 
#>  7 -0.447 -1.27   0.529 -0.429  -2.71   -2.36 
#>  8 -0.167 -0.128 -0.837  1.28    1.96   -1.91 
#>  9  1.39   1.35  -0.130 -1.55   -1.95    4.09 
#> 10  0.876  1.16  -0.158  1.50    2.68    1.13 
#> # … with 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> # A tibble: 100 × 6
#>      cd45   cd38   cd34   cd19 .umap1 .umap2
#>     <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 -0.195 -0.482  0.943  0.475  0.686 -2.43 
#>  2  0.355  0.197 -0.411  0.426  1.79  -1.77 
#>  3  1.67   0.477  1.14   0.623 -0.398 -4.33 
#>  4 -0.212 -2.68   0.392  0.421 -1.19  -1.22 
#>  5 -0.221  0.799  0.184  0.581  2.36  -2.14 
#>  6 -1.03   0.553 -0.847  0.410  1.84  -0.904
#>  7 -0.447 -1.27   0.529 -0.429 -1.28  -1.77 
#>  8 -0.167 -0.128 -0.837  1.28   1.17   0.959
#>  9  1.39   1.35  -0.130 -1.55   4.25  -3.54 
#> 10  0.876  1.16  -0.158  1.50   3.05  -2.84 
#> # … with 90 more rows