Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>      cd45    cd38    cd34    cd19   .pc1   .pc2     .pc3    .pc4
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>    <dbl>   <dbl>
#>  1 -1.05   0.0134  0.369   0.619   0.142 -0.593 -0.785    0.625 
#>  2 -1.37  -1.70   -0.0573  1.26    0.770 -2.05   0.00326  0.643 
#>  3  0.978 -1.70   -0.596   2.57    2.28  -0.596  1.82     0.822 
#>  4 -1.41  -1.95    2.14    0.852   1.69  -2.03  -1.62    -0.448 
#>  5  0.154 -0.477  -2.54   -2.90   -3.08  -0.153  1.68    -1.66  
#>  6 -0.408  0.863   0.0203 -0.649  -0.904  0.467 -0.638   -0.0863
#>  7  0.137 -0.828  -1.22   -0.508  -0.697 -0.435  1.13    -0.501 
#>  8 -0.793  0.289   1.13    0.0218  0.144 -0.129 -1.38    -0.0422
#>  9  1.02  -0.756   1.66    0.827   1.99   0.380 -0.374   -0.801 
#> 10  0.445 -0.760   0.460   1.26    1.44  -0.144  0.270    0.131 
#> # ℹ 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>      cd45    cd38    cd34    cd19 .tsne1  .tsne2
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>
#>  1 -1.05   0.0134  0.369   0.619   2.10   0.0188
#>  2 -1.37  -1.70   -0.0573  1.26    3.04  -2.53  
#>  3  0.978 -1.70   -0.596   2.57   -3.04  -2.93  
#>  4 -1.41  -1.95    2.14    0.852   3.97  -1.76  
#>  5  0.154 -0.477  -2.54   -2.90   -2.19   3.38  
#>  6 -0.408  0.863   0.0203 -0.649   1.36   2.71  
#>  7  0.137 -0.828  -1.22   -0.508  -2.51   0.400 
#>  8 -0.793  0.289   1.13    0.0218  2.76   0.118 
#>  9  1.02  -0.756   1.66    0.827   1.48  -3.15  
#> 10  0.445 -0.760   0.460   1.26    0.506 -2.07  
#> # ℹ 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> # A tibble: 100 × 6
#>      cd45    cd38    cd34    cd19 .umap1 .umap2
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>
#>  1 -1.05   0.0134  0.369   0.619  -3.10  -1.51 
#>  2 -1.37  -1.70   -0.0573  1.26   -1.47   1.44 
#>  3  0.978 -1.70   -0.596   2.57   -0.227  2.16 
#>  4 -1.41  -1.95    2.14    0.852  -3.44  -2.42 
#>  5  0.154 -0.477  -2.54   -2.90    1.92  -1.13 
#>  6 -0.408  0.863   0.0203 -0.649   0.550 -1.20 
#>  7  0.137 -0.828  -1.22   -0.508   1.49   0.728
#>  8 -0.793  0.289   1.13    0.0218 -2.41  -1.82 
#>  9  1.02  -0.756   1.66    0.827  -2.19  -2.77 
#> 10  0.445 -0.760   0.460   1.26    0.221  0.840
#> # ℹ 90 more rows