Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>       cd45   cd38     cd34   cd19   .pc1   .pc2   .pc3    .pc4
#>      <dbl>  <dbl>    <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1 -0.174   0.348 -0.273    1.38  -1.30  -0.116  0.143  0.735 
#>  2  0.0323 -0.422 -0.820   -0.193  0.294  0.504 -0.682  0.276 
#>  3  0.245  -0.668  0.00640  0.173  0.395 -0.335 -0.163  0.558 
#>  4 -0.488  -1.87   0.460   -1.66   2.66   0.347  0.435  0.0317
#>  5  0.154  -0.804 -0.435    0.945 -0.158 -0.115 -0.280  1.30  
#>  6  1.02   -0.144 -0.372   -0.378  0.369 -0.544 -1.15  -0.151 
#>  7  1.07   -0.231 -0.107   -0.348  0.469 -0.781 -0.983 -0.118 
#>  8 -0.0284  1.07  -0.374   -0.313 -0.626  0.207 -0.360 -0.999 
#>  9 -1.34    0.657 -0.741    0.228 -0.861  1.37   0.476 -0.235 
#> 10 -0.530   1.19  -1.24     1.56  -2.28   0.791 -0.287  0.436 
#> # ℹ 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>       cd45   cd38     cd34   cd19 .tsne1  .tsne2
#>      <dbl>  <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1 -0.174   0.348 -0.273    1.38  -2.88   1.61  
#>  2  0.0323 -0.422 -0.820   -0.193  0.653 -0.0307
#>  3  0.245  -0.668  0.00640  0.173  0.718 -0.289 
#>  4 -0.488  -1.87   0.460   -1.66   2.79  -2.92  
#>  5  0.154  -0.804 -0.435    0.945 -0.702 -0.746 
#>  6  1.02   -0.144 -0.372   -0.378  1.97   0.899 
#>  7  1.07   -0.231 -0.107   -0.348  2.03   0.959 
#>  8 -0.0284  1.07  -0.374   -0.313 -1.65   2.00  
#>  9 -1.34    0.657 -0.741    0.228 -3.66  -0.436 
#> 10 -0.530   1.19  -1.24     1.56  -3.80   1.23  
#> # ℹ 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> # A tibble: 100 × 6
#>       cd45   cd38     cd34   cd19 .umap1 .umap2
#>      <dbl>  <dbl>    <dbl>  <dbl>  <dbl>  <dbl>
#>  1 -0.174   0.348 -0.273    1.38  -1.29   0.992
#>  2  0.0323 -0.422 -0.820   -0.193  0.638  2.54 
#>  3  0.245  -0.668  0.00640  0.173 -0.652  0.298
#>  4 -0.488  -1.87   0.460   -1.66   2.02  -0.812
#>  5  0.154  -0.804 -0.435    0.945 -0.688  0.688
#>  6  1.02   -0.144 -0.372   -0.378  0.923  3.51 
#>  7  1.07   -0.231 -0.107   -0.348  0.937  3.45 
#>  8 -0.0284  1.07  -0.374   -0.313 -0.594 -2.63 
#>  9 -1.34    0.657 -0.741    0.228 -2.04  -3.53 
#> 10 -0.530   1.19  -1.24     1.56  -1.90   1.20 
#> # ℹ 90 more rows