Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>       cd45    cd38   cd34    cd19    .pc1    .pc2   .pc3    .pc4
#>      <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>
#>  1  0.767  -0.204   1.94   0.590  -1.41    1.65    0.892  0.654 
#>  2  0.606   0.113  -0.631  2.53   -1.69   -0.735  -1.25   1.03  
#>  3  0.591   0.842  -0.155 -0.581  -0.418   0.0580  0.219 -1.10  
#>  4  0.745   0.346   0.861 -1.31   -0.178   1.22    0.864 -1.05  
#>  5 -1.27    0.911  -0.526  0.341  -0.0431 -1.48    0.759  0.104 
#>  6  0.107   0.143  -0.528  0.358  -0.250  -0.348  -0.295 -0.0260
#>  7 -0.867   2.47   -0.591 -1.92    0.309  -1.45    1.69  -2.33  
#>  8  0.0912 -0.764   0.461 -0.296   0.247   0.895   0.252  0.377 
#>  9  1.33   -0.718   1.70  -0.169  -0.851   2.27    0.423  0.205 
#> 10 -0.687  -0.0813 -1.01   0.0561  0.571  -0.882  -0.198  0.147 
#> # ℹ 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>       cd45    cd38   cd34    cd19 .tsne1 .tsne2
#>      <dbl>   <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#>  1  0.767  -0.204   1.94   0.590  -3.79  -0.826
#>  2  0.606   0.113  -0.631  2.53    1.62   4.63 
#>  3  0.591   0.842  -0.155 -0.581  -1.29  -0.563
#>  4  0.745   0.346   0.861 -1.31   -2.20  -1.53 
#>  5 -1.27    0.911  -0.526  0.341  -0.905  2.22 
#>  6  0.107   0.143  -0.528  0.358   0.666  1.49 
#>  7 -0.867   2.47   -0.591 -1.92   -3.60   2.04 
#>  8  0.0912 -0.764   0.461 -0.296   1.08  -1.54 
#>  9  1.33   -0.718   1.70  -0.169  -3.38  -1.56 
#> 10 -0.687  -0.0813 -1.01   0.0561  0.475  1.01 
#> # ℹ 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> # A tibble: 100 × 6
#>       cd45    cd38   cd34    cd19 .umap1 .umap2
#>      <dbl>   <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#>  1  0.767  -0.204   1.94   0.590  -1.82  -2.34 
#>  2  0.606   0.113  -0.631  2.53    5.40  -1.76 
#>  3  0.591   0.842  -0.155 -0.581  -0.862 -0.984
#>  4  0.745   0.346   0.861 -1.31   -0.251 -1.45 
#>  5 -1.27    0.911  -0.526  0.341   2.88   0.275
#>  6  0.107   0.143  -0.528  0.358   3.45   0.877
#>  7 -0.867   2.47   -0.591 -1.92   -2.02  -0.788
#>  8  0.0912 -0.764   0.461 -0.296  -2.45   2.43 
#>  9  1.33   -0.718   1.70  -0.169  -0.736 -2.08 
#> 10 -0.687  -0.0813 -1.01   0.0561  3.23   1.46 
#> # ℹ 90 more rows