Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>      cd45    cd38   cd34    cd19   .pc1    .pc2    .pc3    .pc4
#>     <dbl>   <dbl>  <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>
#>  1 -1.10  -0.127  -1.74   1.44    0.133 -1.81   -0.0753  1.34  
#>  2  0.530 -0.377  -1.16   0.889   0.439 -0.379  -1.08    0.751 
#>  3  1.83  -1.75   -0.359  0.641   1.58   0.324  -2.24   -0.326 
#>  4 -0.637  1.05   -0.765 -0.469  -1.25  -0.0862  0.423   0.0991
#>  5 -1.84  -0.385   0.524  0.214   0.415 -1.14    1.35   -0.909 
#>  6 -1.18   0.498  -1.74   0.0575 -1.19  -1.32    0.0733  0.615 
#>  7  1.41   0.829   1.30  -1.26   -0.290  2.42   -0.0800 -1.32  
#>  8 -0.671  0.0307  1.08  -0.114   0.460  0.233   1.01   -1.05  
#>  9  0.292  1.92   -0.785 -0.0763 -1.25   0.924   0.259   0.969 
#> 10  0.479  1.53   -1.09   0.231  -0.960  0.659  -0.183   1.18  
#> # ℹ 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>      cd45    cd38   cd34    cd19 .tsne1  .tsne2
#>     <dbl>   <dbl>  <dbl>   <dbl>  <dbl>   <dbl>
#>  1 -1.10  -0.127  -1.74   1.44    3.61   1.29  
#>  2  0.530 -0.377  -1.16   0.889   1.85  -0.949 
#>  3  1.83  -1.75   -0.359  0.641   0.383 -3.90  
#>  4 -0.637  1.05   -0.765 -0.469  -2.60   0.944 
#>  5 -1.84  -0.385   0.524  0.214   0.449  3.32  
#>  6 -1.18   0.498  -1.74   0.0575  4.65   0.978 
#>  7  1.41   0.829   1.30  -1.26   -4.73   2.57  
#>  8 -0.671  0.0307  1.08  -0.114  -0.826  2.54  
#>  9  0.292  1.92   -0.785 -0.0763 -5.07   0.0356
#> 10  0.479  1.53   -1.09   0.231  -5.21  -0.510 
#> # ℹ 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> # A tibble: 100 × 6
#>      cd45    cd38   cd34    cd19 .umap1 .umap2
#>     <dbl>   <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#>  1 -1.10  -0.127  -1.74   1.44    1.16   1.84 
#>  2  0.530 -0.377  -1.16   0.889   1.58   0.429
#>  3  1.83  -1.75   -0.359  0.641   1.16  -0.766
#>  4 -0.637  1.05   -0.765 -0.469  -2.62   0.913
#>  5 -1.84  -0.385   0.524  0.214   4.08  -0.167
#>  6 -1.18   0.498  -1.74   0.0575 -0.516  1.46 
#>  7  1.41   0.829   1.30  -1.26   -3.58  -1.04 
#>  8 -0.671  0.0307  1.08  -0.114   4.42  -0.990
#>  9  0.292  1.92   -0.785 -0.0763 -4.92   1.19 
#> 10  0.479  1.53   -1.09   0.231  -5.11   1.20 
#> # ℹ 90 more rows