Apply dimensionality reduction to a single-cell dataset.

This function is a wrapper around tidytof's tof_reduce_* function family. It performs dimensionality reduction on single-cell data using a user-specified method (of 3 choices) and each method's corresponding input parameters

Usage

tof_reduce_dimensions(
  tof_tibble,
  ...,
  augment = TRUE,
  method = c("pca", "tsne", "umap")
)

Arguments

tof_tibble

A `tof_tbl` or `tibble`.

...

Arguments to be passed to the tof_reduce_* function corresponding to the embedding method. See tof_reduce_pca, tof_reduce_tsne, and tof_reduce_umap.

augment

A boolean value indicating if the output should column-bind the dimensionality-reduced embedding vectors of each cell as a new column in `tof_tibble` (TRUE, the default) or if a tibble including only the low-dimensionality embeddings should be returned (FALSE).

method

A method of dimensionality reduction. Currently, PCA, tSNE, and UMAP embedding are supported.

Value

A tibble with the same number of rows as `tof_tibble`, each representing a single cell. Each of the `num_comp` columns represents each cell's embedding in the calculated embedding space.

See also

Examples

# simulate single-cell data
sim_data <-
    dplyr::tibble(
        cd45 = rnorm(n = 100),
        cd38 = rnorm(n = 100),
        cd34 = rnorm(n = 100),
        cd19 = rnorm(n = 100)
    )

# calculate pca
tof_reduce_dimensions(tof_tibble = sim_data, method = "pca")
#> # A tibble: 100 × 8
#>      cd45   cd38    cd34   cd19   .pc1   .pc2   .pc3   .pc4
#>     <dbl>  <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1  0.934 -0.878  1.18   -0.942 -0.712  1.60  -0.420 -0.493
#>  2 -0.731  1.84  -0.588   0.921  0.497 -2.01  -0.765  0.547
#>  3  0.511  0.536 -0.129   1.34   0.346 -1.15  -0.416 -0.995
#>  4  0.453 -2.25   0.441   0.561  0.756  1.30   0.866 -1.42 
#>  5  1.85  -0.780  0.0438 -0.467 -1.28   0.651  0.334 -1.17 
#>  6  0.112  0.489 -0.640  -0.627 -0.675 -0.446  0.172  0.565
#>  7 -1.13  -0.780 -0.739   0.455  1.14  -0.238  0.956  0.325
#>  8  0.565  0.305 -1.08    1.32   0.177 -1.54   0.452 -0.931
#>  9 -0.356  1.05   0.926   1.95   1.37  -1.16  -1.55  -0.834
#> 10  0.987 -1.75   1.55   -1.99  -1.09   2.87  -0.106 -0.199
#> # … with 90 more rows

# calculate tsne
tof_reduce_dimensions(tof_tibble = sim_data, method = "tsne")
#> # A tibble: 100 × 6
#>      cd45   cd38    cd34   cd19 .tsne_1 .tsne_2
#>     <dbl>  <dbl>   <dbl>  <dbl>   <dbl>   <dbl>
#>  1  0.934 -0.878  1.18   -0.942  -1.47    3.91 
#>  2 -0.731  1.84  -0.588   0.921   1.37   -3.06 
#>  3  0.511  0.536 -0.129   1.34    2.19   -0.860
#>  4  0.453 -2.25   0.441   0.561   3.15    2.63 
#>  5  1.85  -0.780  0.0438 -0.467  -0.650   2.74 
#>  6  0.112  0.489 -0.640  -0.627  -1.59    0.116
#>  7 -1.13  -0.780 -0.739   0.455  -0.509  -1.76 
#>  8  0.565  0.305 -1.08    1.32    2.03   -0.862
#>  9 -0.356  1.05   0.926   1.95    3.77   -0.605
#> 10  0.987 -1.75   1.55   -1.99   -2.38    4.11 
#> # … with 90 more rows

# calculate umap
tof_reduce_dimensions(tof_tibble = sim_data, method = "umap")
#> # A tibble: 100 × 6
#>      cd45   cd38    cd34   cd19  .umap1 .umap2
#>     <dbl>  <dbl>   <dbl>  <dbl>   <dbl>  <dbl>
#>  1  0.934 -0.878  1.18   -0.942  2.27    2.16 
#>  2 -0.731  1.84  -0.588   0.921  0.0266 -1.18 
#>  3  0.511  0.536 -0.129   1.34   1.08   -0.886
#>  4  0.453 -2.25   0.441   0.561 -4.32   -1.22 
#>  5  1.85  -0.780  0.0438 -0.467  2.01    1.12 
#>  6  0.112  0.489 -0.640  -0.627  0.210   0.651
#>  7 -1.13  -0.780 -0.739   0.455  0.337  -3.40 
#>  8  0.565  0.305 -1.08    1.32   0.911  -0.452
#>  9 -0.356  1.05   0.926   1.95  -3.20    4.28 
#> 10  0.987 -1.75   1.55   -1.99   2.39    2.27 
#> # … with 90 more rows